% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{add.auto.classification}
\alias{add.auto.classification}
\title{Define polygons for manual population gating.}
\usage{
add.auto.classification(popname, para.x, para.y, position, gates, scale,
  min.pe = NA, gates.log = NULL)
}
\arguments{
\item{popname}{Population name.}

\item{para.x}{Channel to use as x axis.}

\item{para.y}{Channel to use as y axis.}

\item{position}{position parameter to FlowDensity}

\item{gates}{gates parameter to FlowDensity}

\item{scale}{scale parameter to FlowDensity}

\item{min.pe}{Only consider partciesl with pe > min.pe}

\item{gates.log}{Per population named list of classification parameters. If
a definition for popname already exists it will be updated. If it doesn't
exist it will be appended to the end to the list. If NULL, a new list will
be created.}
}
\value{
gates.log with a new FlowDensity func call parameter set for popname
}
\description{
Define polygons for manual population gating.
}
\examples{
\dontrun{
gates.log <- add.auto.classification("beads", "fsc_small", "pe",
                                     c(FALSE,TRUE), c(2.0,NA), 0.975)
gates.log <- add.auto.classification("prochloro", "fsc_small", "chl_small",
                                     c(FALSE,TRUE), c(2.0,NA), 0.975, gates.log)
}
}
