% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{add_abundance}
\alias{add_abundance}
\title{Add volume-normalized abundance to gridded SeaFlow data}
\usage{
add_abundance(gridded, volumes, calib = NULL)
}
\arguments{
\item{gridded}{Gridded data created by grid().}

\item{volumes}{Volume dataframe at the same time resolution as gridded, usually created by
create_volume_table().}

\item{calib}{Optional influx calibration dataframe used to adjust data by population.
Columns should incude pop and a.}
}
\value{
gridded with volume-normalized abundance columns for n_per_uL and Qc_sum_per_uL.
  If calib is provided n and Qc_sum will be adjusted for populations in calib.
}
\description{
Add volume-normalized abundance to gridded SeaFlow data
}
