% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{add_adundance}
\alias{add_adundance}
\title{Add volume-normalized abundance to gridded SeaFlow data}
\usage{
add_adundance(psd, volumes, calib = NULL)
}
\arguments{
\item{psd}{Gridded data created by create_PSD().}

\item{volumes}{Volume dataframe at the same time resolution as psd, usually created by
create_volume_table().}

\item{calib}{Optional influx calibration dataframe used to adjust data by population.
Columns should be pop, a, b.}
}
\value{
psd with volume-normalized abundance columns for n_per_uL and Qc_sum_per_uL.
  If calib is provided n and Qc_sum will be adjusted for populations in calib.
}
\description{
Add volume-normalized abundance to gridded SeaFlow data
}
