% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{add_auto_classification}
\alias{add_auto_classification}
\title{Define polygons for manual population gating.}
\usage{
add_auto_classification(
  popname,
  para_x,
  para_y,
  position,
  gates,
  scale,
  min_pe = NA,
  gates.log = NULL
)
}
\arguments{
\item{popname}{Population name.}

\item{position}{position parameter to FlowDensity}

\item{gates}{gates parameter to FlowDensity}

\item{scale}{scale parameter to FlowDensity}

\item{min_pe}{Only consider particles with pe > min.pe}

\item{gates.log}{Per population named list of classification parameters. If
a definition for popname already exists it will be updated. If it doesn't
exist it will be appended to the end to the list. If NULL, a new list will
be created.}

\item{para.x}{Channel to use as x axis.}

\item{para.y}{Channel to use as y axis.}
}
\value{
gates.log with a new FlowDensity func call parameter set for popname
}
\description{
Define polygons for manual population gating.
}
