% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{add_coord_labels}
\alias{add_coord_labels}
\title{Add string coordinate labels to a dataframe of gridded SeaFlow data}
\usage{
add_coord_labels(gridded, grid_bins)
}
\arguments{
\item{gridded}{Gridded dataframe created by create_gridded.}

\item{grid_bins}{Named list of breakpoints created by create_grid_bins().}
}
\value{
Named list identical to grid with numeric break points converted to character vectors.
}
\description{
Add string coordinate labels to a dataframe of gridded SeaFlow data
}
