% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{add_manual_classification}
\alias{add_manual_classification}
\title{Define polygons for manual population gating.}
\usage{
add_manual_classification(opp, popname, para_x, para_y, gates.log = NULL)
}
\arguments{
\item{opp}{OPP data frame.}

\item{popname}{Population name.}

\item{para_x}{Channel to use as x axis.}

\item{para_y}{Channel to use as y axis.}

\item{gates.log}{Per population named list of classification parameters. If
a definition for popname already exists it will be updated. If it doesn't
exist it will be appended to the end to the list. If NULL, a new list will
be created.}
}
\value{
gates.log with a new polygon defintion for popname.
}
\description{
Define polygons for manual population gating.
}
