% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{add_window_paths}
\alias{add_window_paths}
\title{Add intervals and window paths for each date in df.}
\usage{
add_window_paths(df, window_paths)
}
\arguments{
\item{df}{Dataframe with a "date" column of POSIXct dates.}

\item{window_paths}{Time-windowed SeaFlow file paths.}
}
\value{
A copy of df with a new lubridate::interval column "interval" and a
  a new character column "window_path" of a time-windowed file path matching
  each date. New values are NA if a date doesn't match a time-windowed
  seaflow file path.
}
\description{
Add intervals and window paths for each date in df.
}
