% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{add_window_paths_intervals}
\alias{add_window_paths_intervals}
\title{Get datetime interval for time-windowed SeaFlow file paths.}
\usage{
add_window_paths_intervals(window_df)
}
\arguments{
\item{window_df}{Dataframe with column "window_path" for SeaFlow
time-windowed file paths.}
}
\value{
A copy of window_df with a new "interval" column of
  lubridate::interval, inclusive at the start boundaries and exclusive at
  the end boundary.
}
\description{
Get datetime interval for time-windowed SeaFlow file paths.
}
