% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{auto.classify}
\alias{auto.classify}
\title{Classify particles based on semisupervized clustering method from flowDensity package.}
\usage{
auto.classify(opp, params, popname)
}
\arguments{
\item{opp}{OPP data frame.}

\item{params}{Named list of gating parameters. Must contain a params$poly
entry with polygon definitions.}

\item{popname}{Name of the population}
}
\value{
List of per particle classifications.
}
\description{
Classify particles based on semisupervized clustering method from flowDensity package.
}
\examples{
\dontrun{
vct <- auto.classify(opp, params, popname)
}
}
