% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{carbon_conversion}
\alias{carbon_conversion}
\title{Estimate Carbon cell quotas and biomass based on normalized foward scatter to 1-µm beads used as internal standard}
\usage{
carbon_conversion(stat, inst = NULL)
}
\arguments{
\item{stat}{Table generated by the function normalization().}

\item{inst}{Instrument serial. If not provided this will attempt to be
parsed from the SFL file name (<cruise>_<serial>.sfl).}
}
\value{
A dataframe with carbon cell quotas and biomass
}
\description{
Estimate Carbon cell quotas and biomass based on normalized foward scatter to 1-µm beads used as internal standard
}
