% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{check.for.populated.sfl}
\alias{check.for.populated.sfl}
\title{Check if SFL table is populated.}
\usage{
check.for.populated.sfl(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
None
}
\description{
Check if SFL table is populated.
}
\examples{
\dontrun{
check.for.populated.sfl(db)
}
}

