% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{classify.opp.files}
\alias{classify.opp.files}
\title{Classifiy particles for a list of OPP files.}
\usage{
classify.opp.files(db, cruise.name, opp.dir, opp.files, vct.dir,
  gating.id = NULL)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{cruise.name}{Cruise name.}

\item{opp.dir}{OPP file directory.}

\item{opp.files}{List of OPP files to classify. Include julian day directory.}

\item{vct.dir}{VCT file output directory.}

\item{gating.id}{Optionally provide the ID for gating parameters. If NULL,
the most recently saved gating parameters will be used.}
}
\value{
None
}
\description{
Classify a  list of OPP files. Save per file aggregate population statistics
to SQLite3 database and save particle population definitions to text files
in vct.dir.
}
\examples{
\dontrun{
classify.opp.files(db, "testcruise", opp.dir, opp.files, vct.dir)
classify.opp.files(db, "testcruise", opp.dir, opp.files, vct.dir,
                   "d3afb1ea-ad20-46cf-866d-869300fe17f4")
}
}
