% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{classify.opp}
\alias{classify.opp}
\title{Classify particles with a generic gating function.}
\usage{
classify.opp(opp, classify.func, ...)
}
\arguments{
\item{opp}{OPP data frame.}

\item{classify.func}{Classifying function.}
}
\value{
List of per particle classifications.
}
\description{
Classify particles with a generic gating function.
}
\examples{
\dontrun{
vct <- classify.opp(opp, ManualGating, poly.log)
}
}

