% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{classify_opp}
\alias{classify_opp}
\title{Classify particles from an OPP data frame.}
\usage{
classify_opp(opp, gates.log)
}
\arguments{
\item{opp}{Data frame of linearized filtered particles.}

\item{gates.log}{A gating scheme from the function "add_manual_classification()"
or "add_auto_classification()"}
}
\value{
List of per particle classifications
}
\description{
Classify particles from an OPP data frame using a gating scheme provided by gates.log.
}
