% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{classify_opp_files}
\alias{classify_opp_files}
\title{Classify particles for a list of OPP files.}
\usage{
classify_opp_files(
  db,
  opp_dir,
  opp_files,
  vct_dir,
  gating_id = NULL,
  mie_table = NULL,
  cores = 1,
  filter_id_check = TRUE
)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{opp_dir}{OPP file directory.}

\item{opp_files}{List of OPP files to classify. Include julian day directory.
If NULL then all OPP files (regardless of outlier status) will attempt to
be classified.}

\item{vct_dir}{VCT file output directory.}

\item{gating_id}{Single gating ID to use for all files in opp_files.}

\item{mie_table}{Optionally provide the Mie theory table. If NULL,
the installed Mie Theory lookup table will be used. This table should be
identical to one produced by read_mie_csv().}

\item{cores}{Number of cores to use for gating}

\item{filter_id_check}{Set to FALSE to disable check that filter_id in OPP
file matches filter_id from the database for the given file_id.}
}
\value{
None
}
\description{
Classify a  list of OPP files. Save per file aggregate population statistics
to SQLite3 database and save particle population definitions to text files
in vct_dir.
}
