% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{clean_file_path}
\alias{clean_file_path}
\title{Clean a file path.}
\usage{
clean_file_path(paths)
}
\arguments{
\item{paths}{Character vector of file paths to clean.}
}
\value{
Modified file path as julian_day/EVT_file_name.
}
\description{
Convert an EVT/OPP/VCT file path to a form suitable for storage in the SQLite
db. Any ".gz", ".opp", ".vct" extensions will be removed.
}
