% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{combine.quantile}
\alias{combine.quantile}
\title{Calculate mean and sd from the different quantiles.}
\usage{
combine.quantile(stat)
}
\arguments{
\item{stat}{Table generated by the function size.carbon.conversion.}
}
\value{
Data frame of aggregate statistics.
}
\description{
Calculate mean and sd from the different quantiles.
}
