% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\name{concatenate.evt}
\alias{concatenate.evt}
\title{Concatenate EVT files}
\usage{
concatenate.evt(evt.list, evt.dir, n = 1e+05, min.fsc = 0,
  min.pe = 0, min.chl = 0, transform = TRUE, ...)
}
\arguments{
\item{evt.list}{List of EVT files.}

\item{evt.dir}{Path to EVT files.}

\item{n}{Number of rows to return.}

\item{min.fsc, }{min.pe, min.chl Minimum value for fsc_small, pe and chl_small respectively}
}
\value{
A dataframe with n rows.
}
\description{
Concatenate EVT files
}
