% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\name{concatenate.opp}
\alias{concatenate.opp}
\title{Concatenate OPP files}
\usage{
concatenate.opp(
  opp.list,
  opp.dir,
  n = 1e+05,
  min.fsc = 0,
  min.pe = 0,
  min.chl = 0,
  transform = TRUE,
  ...
)
}
\arguments{
\item{opp.list}{List of OPP files.}

\item{opp.dir}{Path to OPP files.}

\item{n}{Number of rows to return.}

\item{min.fsc, }{min.pe, min.chl Minimum value for fsc_small, pe and chl_small respectively}
}
\value{
A dataframe with n rows.
}
\description{
Concatenate OPP files
}
