% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{convert.opp.files}
\alias{convert.opp.files}
\title{Convert fsc_small values to cell diameter (diam, µm) and carbon quotas (Qc, pg)}
\usage{
convert.opp.files(db, opp.dir, opp.files, vct.dir)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{opp.dir}{OPP file directory.}

\item{opp.files}{List of OPP files to classify. Include julian day directory.}

\item{vct.dir}{VCT file output directory.}
}
\value{
Cell diameter (diam, µm) and carbon quotas (Qc, pg) for each particle
}
\description{
Convert fsc_small values to cell diameter (diam, µm) and carbon quotas (Qc, pg). Save per particle values
to SQLite3 database in vct.dir.
}
\examples{
\dontrun{
convert.opp.files(db, opp.dir, opp.files, vct.dir)
}
}
