% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{copy_outlier_table}
\alias{copy_outlier_table}
\title{Copy outlier entries from src_db to dest_db.}
\usage{
copy_outlier_table(db_from, db_to)
}
\arguments{
\item{db_from}{Popcycle database to copy flags > 0 from.}

\item{db_to}{Popcycle database to copy flags > 0 to.}
}
\value{
None
}
\description{
This assumes db_to already has a complete outlier table with entries for
every file in the opp table. Entries in db_to with flag values in db_from
> 0 will be updated to reflect db_from. db_from entries without a
corresponding entry in db_to will be ignored.
}
