% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{copy_tables}
\alias{copy_tables}
\title{Copy tables from one popcycle database to another.}
\usage{
copy_tables(db_from, db_to, tables)
}
\arguments{
\item{db_from}{Popcycle database to copy tables from.}

\item{db_to}{Popcycle database to copy tables to.}
}
\value{
None
}
\description{
The source database will remained unchanged. The destination database will
have selected table cleared before the copy. Schemas for source and
destination tables must match or an error is thrown.
}
\examples{
\dontrun{
copy_tables(db_from, db_to)
}
}
