% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{create.filter.params}
\alias{create.filter.params}
\title{Get Notch and Offset values for Filter.notch function.}
\usage{
create.filter.params(inst, fsc, d1, d2)
}
\arguments{
\item{inst}{Instrument serial number}

\item{fsc}{Small forward scatter value of 1 µm beads}

\item{d1}{D1 value of 1 µm beads}

\item{d2}{D2 value of 1 µm beads}
}
\value{
Data frame with filtering parameters for 2.5, 50.0, 97.5 quantiles
}
\description{
Get Notch and Offset values for Filter.notch function.
}
\examples{
\dontrun{
filt <- create.filter.params(inst, fsc, d1, d2)
}
}
