% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{create_breaks}
\alias{create_breaks}
\title{Create log-spaced breaks vector that can be used with cut().}
\usage{
create_breaks(bins, minval, maxval, log_base = 2)
}
\arguments{
\item{bins}{Number of bins in distribution.}

\item{minval}{Value for the minimum break.}

\item{maxval}{Value for the maximum break.}

\item{log_base}{Log base to use when creating log-spaced bins for distribution.}
}
\value{
Vector defining breaks for use with cut().
}
\description{
Create log-spaced breaks vector that can be used with cut().
}
\examples{
\dontrun{
breaks <- create_breaks(40, 1, 3500, 10)
}
}
