% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{create_breaks2}
\alias{create_breaks2}
\title{Create breaks vector that can be used with cut(), with log2 bin width as input.}
\usage{
create_breaks2(bin_count, min_val, log2_bin_width_inv, log_answers = TRUE)
}
\arguments{
\item{bin_count}{Number of bins between breaks.}

\item{min_val}{Value for the minimum break (fencepost).}

\item{log2_bin_width_inv}{1 / (log2(bin width)), e.g. 8 means a log2(bin_width) of 0.125}

\item{log_answers}{This determines whether the break values themselves are log
values or not. e.g. if TRUE breaks may be [1, 2, 3] (log2), if FALSE breaks may be
[2, 4, 8] (linear).}
}
\value{
Vector defining breaks for use with cut(), length == bin_count + 1
}
\description{
Create breaks vector that can be used with cut(), with log2 bin width as input.
}
