% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{create_filter_params}
\alias{create_filter_params}
\title{Construct filter parameters from 1µm bead coordinates}
\usage{
create_filter_params(
  inst,
  fsc,
  d1,
  d2,
  min_d1,
  min_d2,
  width = 5000,
  slope_file = NULL
)
}
\arguments{
\item{inst}{Instrument serial}

\item{fsc}{25th, 50, 75th quantiles of bead fsc_small}

\item{d1}{25th, 50, 75th quantiles of bead D1}

\item{d2}{25th, 50, 75th quantiles of bead D2}

\item{min_d1}{Y-offset for small particle filter in D1}

\item{min_d2}{Y-offset for small particle filter in D2}

\item{width}{Tolerance for D1 == D2 alignment test during filtering}

\item{slope_file}{CSV file of instrument-calibrated filtering slopes}
}
\value{
Data frame of filtering parameters
}
\description{
Construct filter parameters from 1µm bead coordinates
}
