% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{create_grid_bins}
\alias{create_grid_bins}
\title{Create the breaks to use to grid VCT data.}
\usage{
create_grid_bins(
  bin_count = 85,
  channel_range = c(1, 3200),
  diam_range = c(0.1, 37),
  qc_min_val = 0.002,
  qc_log2_bin_width_inv = 8,
  log_base = NULL,
  log_answers = TRUE
)
}
\arguments{
\item{bin_count}{Number of bins between breaks.}

\item{channel_range}{Vector of range to use for standard SeaFlow channel breaks.}

\item{diam_range}{Vector of range to use for diam breaks.
#' @param log_base Log base to use when creating log-spaced bins for distribution.
  If NULL the breaks will not be log-spaced.}

\item{qc_min_val}{Minimum value for Qc.}

\item{qc_log2_bin_width_inv}{Qc bin width as 1 / log2(bin_width), e.g. a value
of 8 means a log2(bin_width) of .125.}

\item{log_answers}{If log_base is provided, this determines whether the break
values themselves are log values or not. e.g. if TRUE breaks may be
[1, 2, 3], if FALSE breaks may be [10, 100, 1000].}
}
\value{
Named list defining breaks for use with cut(). Each item has length == bins + 1.
}
\description{
Create the breaks to use to grid VCT data.
}
