% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{create_meta}
\alias{create_meta}
\title{Create a metadata tibble for one quantile from the SFL table.}
\usage{
create_meta(db, quantile)
}
\arguments{
\item{db}{popcycle database file.}

\item{quantile}{OPP filtering quantile to use.}
}
\value{
A tibble of date, volume, opp_evt_ratio, flag
}
\description{
Create a metadata tibble for one quantile from the SFL table.
}
