% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{create_PSD}
\alias{create_PSD}
\title{Create particle size distributions as particle counts and Qc sums for many VCT files}
\usage{
create_PSD(
  vct_files,
  quantile,
  refracs,
  grid,
  log_base = NULL,
  max_boundary_proportion = NULL,
  ignore_dates = NULL,
  pop = NULL,
  use_data.table = TRUE,
  cores = 1
)
}
\arguments{
\item{vct_files}{VCT files to grid.}

\item{quantile}{OPP filtering quantile.}

\item{refracs}{Dataframe of refractive indices to use for each population. Column names
should be population names used in classification, values should be mid, lwr, upr.}

\item{grid}{Named list of break points to use for gridding. Names should be fsc_small, pe,
chl_small, Qc, diam.}

\item{log_base}{If the break points are logged values, provide the base here to properly
log the VCT data before gridding. If break points are not log use NULL.}

\item{max_boundary_proportion}{Proportion of particles in each VCT file which can be
removed as boundary points before discarding all data in the file.
Should be > 0 and <= 1. If this parameter is NULL, no boundary points will be removed
before gridding.}

\item{ignore_dates}{Don't process VCT data with these dates.}

\item{pop}{A single populations label to keep. Particles not matching this label will be
ignored.}

\item{use_data.table}{Use data.table for performance speedup, otherwise use dplyr.}

\item{cores}{Number of cores to use}
}
\value{
Tibble of gridded data, grouped by date, fsc_small, pe, chl_small, Qc diam grid
  locations, and population. Each grid coordinate is an integer index into the corresponding
  grid breaks vector. The break point at that location defines the inclusive lower bound
  of the bin, and the next break point in the sequence defines the exclusive upper bound.
  Data columns summarizing each group are n for particle count and Qc_sum for the the sum
  of Qc.
}
\description{
Create particle size distributions as particle counts and Qc sums for many VCT files
}
