% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{create_PSD_one_file}
\alias{create_PSD_one_file}
\title{Create particle size distribution as raw counts for one VCT file}
\usage{
create_PSD_one_file(
  vct_file,
  data_col,
  quantile,
  bins,
  minval,
  maxval,
  maxval_headroom = 1e-06,
  log_base = 2
)
}
\arguments{
\item{vct_file}{VCT file path.}

\item{data_col}{VCT column to use as size metric: fsc_small, chl_small, pe, Qc, or diam. If
Qc or diam is selected, the refractive index is indicated in the extra column "n".}

\item{quantile}{OPP Filtering quantile.}

\item{bins}{Number of bins in distribution.}

\item{minval}{Value of smallest bin. Exclusive so set lower than lowest expected value.}

\item{maxval}{Value of largest bin. Inclusive. To account for floating point imprecision
a small value of maxval_headroom will be added to maxval automatically.}

\item{maxval_headroom}{Amount to add to maxval to account for floating point imprecision.}

\item{log_base}{Log base to use when creating log-spaced bins for distribution.}
}
\value{
Size distribution dataframe.
}
\description{
Create particle size distribution as raw counts for one VCT file
}
\examples{
\dontrun{
psd <- create_PSD_one_file(vct_file, "fsc_small", quantile = 50, bins = 40, minval = 1, maxval = 3500)
}
}
