% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{date.to.db.date}
\alias{date.to.db.date}
\title{Convert a date string to format suitable for popcycle db date comparison.}
\usage{
date.to.db.date(date.string)
}
\arguments{
\item{date.string}{Date text in format YYYY-MM-DD HH:MM.}
}
\value{
Date text as YYYY-MM-DDTHH:MM:00.
}
\description{
Convert a date string to format suitable for popcycle db date comparison.
}
\examples{
\dontrun{
db.date <- date.to.db.date("2014-01-24 13:03")
}
}
