% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{delete.filter.params.by.id}
\alias{delete.filter.params.by.id}
\title{Delete DB filter parameters by ID.}
\usage{
delete.filter.params.by.id(db, filter.id)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{filter.id}{ID for filter entries.}
}
\value{
None
}
\description{
Delete DB filter parameters by ID.
}
\examples{
\dontrun{
delete.filter.params.by.id(db, "d3afb1ea-ad20-46cf-866d-869300fe17f4")
}
}
