% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{delete.gating.params.by.id}
\alias{delete.gating.params.by.id}
\title{Delete DB gating parameters by gating ID.}
\usage{
delete.gating.params.by.id(db, gating.id)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{gating.id}{ID for gating and poly entries.}
}
\value{
None
}
\description{
Any gating polygon entries in the poly table will also be deleted.
}
\examples{
\dontrun{
delete.gating.params.by.id(db, "d3afb1ea-ad20-46cf-866d-869300fe17f4")
}
}

