% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{delete.opp.by.file}
\alias{delete.opp.by.file}
\title{Delete an OPP binary file by file name.}
\usage{
delete.opp.by.file(opp.dir, file.name)
}
\arguments{
\item{opp.dir}{OPP file directory.}

\item{file.name}{File name with julian day directory.}
}
\value{
None
}
\description{
Delete an OPP binary file by file name.
}
\examples{
\dontrun{
delete.opp.by.file(opp.dir, "2014_185/2014-07-04T00-00-02+00-00")
}
}
