% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{delete.poly.by.id.pop}
\alias{delete.poly.by.id.pop}
\title{Delete DB poly parameters by gating ID and population}
\usage{
delete.poly.by.id.pop(db, gating.id, popname)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{gating.id}{gating_id for poly entries.}

\item{popname}{Population name}
}
\value{
None
}
\description{
Note: This is usually done via delete.gating.params.by.id.
}
\examples{
\dontrun{
delete.poly.by.id.pop(db, "d3afb1ea-ad20-46cf-866d-869300fe17f4", "beads")
}
}
