% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{delete.poly.by.id}
\alias{delete.poly.by.id}
\title{Delete DB poly parameters by gating ID.}
\usage{
delete.poly.by.id(db, gating.id)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{gating.id}{gating_id for poly entries.}
}
\value{
None
}
\description{
Note: This is usually done via delete.gating.params.by.id.
}
\examples{
\dontrun{
delete.poly.by.id(db, "d3afb1ea-ad20-46cf-866d-869300fe17f4")
}
}

