% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{delete.vct.by.file}
\alias{delete.vct.by.file}
\title{Delete a VCT text file by file name.}
\usage{
delete.vct.by.file(vct.dir, file.name)
}
\arguments{
\item{vct.dir}{VCT file directory.}

\item{file.name}{File name with julian day directory.}
}
\value{
None
}
\description{
Delete a VCT text file by file name.
}
\examples{
\dontrun{
delete.vct.by.file(opp.dir, "2014_185/2014-07-04T00-00-02+00-00")
}
}

