% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{delete_filter_params_by_id}
\alias{delete_filter_params_by_id}
\title{Delete DB filter parameters by ID.}
\usage{
delete_filter_params_by_id(db, filter_id)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{filter_id}{ID for filter entries.}
}
\value{
Number of rows deleted.
}
\description{
Delete DB filter parameters by ID.
}
