% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{delete_gating_params_by_id}
\alias{delete_gating_params_by_id}
\title{Delete DB gating parameters by gating ID.}
\usage{
delete_gating_params_by_id(db, gating_id)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{gating_id}{ID for gating and poly entries.}
}
\value{
Number of rows deleted.
}
\description{
Any gating polygon entries in the poly table will also be deleted.
}
