% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{delete_poly_by_id}
\alias{delete_poly_by_id}
\title{Delete DB poly parameters by gating ID.}
\usage{
delete_poly_by_id(db, gating_id)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{gating_id}{gating_id for poly entries.}
}
\value{
Number of rows deleted.
}
\description{
Note: This is usually done via delete.gating.params.by.id.
}
