% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{delete_poly_by_id_pop}
\alias{delete_poly_by_id_pop}
\title{Delete DB poly parameters by gating ID and population}
\usage{
delete_poly_by_id_pop(db, gating_id, popname)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{popname}{Population name}

\item{gating.id}{gating_id for poly entries.}
}
\value{
Number of rows deleted.
}
\description{
Note: This is usually done via delete.gating.params.by.id.
}
