% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{empty_evt}
\alias{empty_evt}
\title{Create an empty EVT data frame}
\usage{
empty_evt()
}
\value{
EVT data frame with no rows
}
\description{
Create an empty EVT data frame
}
\examples{
\dontrun{
evt <- empty_evt()
}
}
