% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{empty_opp}
\alias{empty_opp}
\title{Create an empty OPP data frame}
\usage{
empty_opp()
}
\value{
OPP data frame with no rows
}
\description{
Create an empty OPP data frame
}
\examples{
\dontrun{
evt <- empty_opp()
}
}
