% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{evaluate.evt}
\alias{evaluate.evt}
\title{Filter and classify one EVT file.}
\usage{
evaluate.evt(db, cruise.name, evt.dir, opp.dir, vct.dir, evt.file)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{cruise.name}{Cruise name.}

\item{evt.dir}{EVT file directory.}

\item{opp.dir}{OPP file output directory.}

\item{vct.dir}{VCT file output directory.}

\item{evt.file}{EVT file to filter. Include julian day directory.}
}
\value{
None
}
\description{
Filter and classify one EVT file.
}
\examples{
\dontrun{
evaluate.evt(db, "testcruise", evt.dir, opp.dir, vct.dir,
             "2014_185/2014-07-04T00-00-02+00-00")
}
}
