% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter.evt.files}
\alias{filter.evt.files}
\title{Filter a list of EVT files.}
\usage{
filter.evt.files(db, cruise.name, evt.dir, evt.files, opp.dir,
  filter.id = NULL)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{cruise.name}{Cruise name.}

\item{evt.dir}{EVT file directory.}

\item{evt.files}{List of EVT files to filter. Include julian day directory.}

\item{opp.dir}{OPP file output directory.}

\item{filter.id}{Optionally provide the ID for filter parameters. If NULL,
the most recently saved filter parameters will be used.}
}
\value{
None
}
\description{
Filter a list of EVT files. Save OPP per file aggregate statistics to
SQLite3 database and save particle data to binary files in opp.dir.
}
\examples{
\dontrun{
filter.evt.files(db, "testcruise", evt.dir, evt.files, opp.dir)
filter.evt.files(db, "testcruise", evt.dir, evt.files, opp.dir,
                 "d3afb1ea-ad20-46cf-866d-869300fe17f4")
}
}
