% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter.evt}
\alias{filter.evt}
\title{Filter EVT particles with a generic filter function.}
\usage{
filter.evt(evt, filter.func, ...)
}
\arguments{
\item{evt}{EVT data frame.}

\item{filter.func}{Filtering function.}
}
\value{
Named list where list$params contains a list of filtering parameters
  and list$opp contains OPP data frame.
}
\description{
Filter EVT particles with a generic filter function.
}
\examples{
\dontrun{
filt <- filter.evt(evt, filter.notch, filter.params)
}
}
