% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_evt}
\alias{filter_evt}
\title{Filter EVT particles.}
\usage{
filter_evt(evt, filter_params)
}
\arguments{
\item{evt}{EVT data frame.}

\item{filter_params}{Filtering parameters in data frame. Should contain
parameters for all quantiles. Columns should include quantile, width,
notch.small.D1, notch.small.D2, notch.large.D1, notch.large.D2,
offset.small.D1, offset.small.D2, offset.large.D1, offset.large.D2.}
}
\value{
OPP data frame with new logical columns for each quantile.
}
\description{
Filter EVT particles.
}
