% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_evt_files}
\alias{filter_evt_files}
\title{Filter a list of EVT files.}
\usage{
filter_evt_files(
  db,
  evt_dir,
  evt_files,
  opp_dir,
  filter_id = NULL,
  max_particles_per_file = MAX_PARTICLES_PER_FILE_DEFAULT,
  max_opp_evt_ratio = MAX_OPP_EVT_RATIO_DEFAULT,
  cores = 1
)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{evt_dir}{EVT file directory.}

\item{evt_files}{List of EVT file IDs to filter. Include julian day directory.
Pass NULL to filter all EVT files discovered in evt_dir.}

\item{opp_dir}{OPP file output directory.}

\item{filter_id}{Optionally provide the ID for filter parameters. Pass NULL
to filter using schedule describedin the filter_plan table.}

\item{max_particles_per_file}{Maximum number of particles per 3 minute EVT file.
Set to NULL to disable.}

\item{max_opp_evt_ratio}{Maximum OPP/EVT ratio to accept. Set to NULL to disable.}

\item{cores}{Number of logical cores to use.}
}
\value{
None
}
\description{
Filter a list of EVT files. Save OPP per file aggregate statistics to
SQLite3 database and save particle data to hourly parquet files in opp_dir.
Set outlier flag to 0 for all files without a current entry in the outlier
table, otherwise leave the outlier entry as is.
}
