% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{find.filter.notch}
\alias{find.filter.notch}
\title{find.filter.notch}
\usage{
find.filter.notch(evt.list, origin = NA, width = 0.5, notch = c(NA, NA),
  offset = 0, do.plot = TRUE)
}
\description{
find.filter.notch
}

