% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{find_common_dbs}
\alias{find_common_dbs}
\title{Find common database files between two directories.}
\usage{
find_common_dbs(dir_a, dir_b)
}
\arguments{
\item{dir_a, }{dir_b Directories to compare.}
}
\value{
Data Frame with columns for basename, old_path, and new_path
}
\description{
Directories will be search recursively and files will be matched by
basename. An error will be thrown if the same database file occurs more than
}
\examples{
\dontrun{
common <- find_common_dbs(dir_a, dir_b)
}
}
