% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{from_filename_date_str}
\alias{from_filename_date_str}
\title{Convert timestamp component of SeaFlow filenames to POSIXct.}
\usage{
from_filename_date_str(timestamps)
}
\arguments{
\item{timestamps}{Character vector of timestamp components of SeaFlow
filenames matching the format YYYY-MM-DDTHH-MM-SS+HH-MM, for example
"2020-07-23T14-04-05+00-00". This corresponds to the RFC3339 formatted
timestamp "2020-07-23T14:04:05+00:00".}
}
\value{
POSIXct vector
}
\description{
Convert timestamp component of SeaFlow filenames to POSIXct.
}
