% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get.clean.stat.table}
\alias{get.clean.stat.table}
\title{Clean the stat table by selecting the closest refractive index to the reference diameter for a particular population,
keep only the clean data (flag ==0 ) and remove percentile column}
\usage{
get.clean.stat.table(db, pop = "prochloro", ref_diam = 0.54)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{pop}{Name of the population. Can be either "prochloro", "synecho", "picoeuk", "croco"}

\item{ref_diam}{Diameter value (in micron)}
}
\value{
A cleaned stat table with corrected refractive index
}
\description{
Clean the stat table by selecting the closest refractive index to the reference diameter for a particular population,
keep only the clean data (flag ==0 ) and remove percentile column
}
