% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.cruise}
\alias{get.cruise}
\title{Get cruise name}
\usage{
get.cruise(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Cruise name
}
\description{
Get cruise name
}
\examples{
\dontrun{
cruise <- get.cruise(db)
}
}
