% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.cytdiv.table}
\alias{get.cytdiv.table}
\title{Get cytometric diversity table cytdiv.}
\usage{
get.cytdiv.table(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Data frame of cytometric diversity table.
}
\description{
Get cytometric diversity table cytdiv.
}
\examples{
\dontrun{
cytdiv.table <- get.cytdiv.table(db)
}
}

