% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get.dawn.dusk.time}
\alias{get.dawn.dusk.time}
\title{Extract dawn and dusk times from a time series of PAR values}
\usage{
get.dawn.dusk.time(x, cutoff)
}
\arguments{
\item{x}{A data frame with two columns. The first column must be time values
in as.POSIXct format. The second column must be PAR values.}

\item{cutoff}{An integer representing the smallest par value that is considered
daytime.}
}
\value{
A vector consisting of dawn and dusk time values in as.POSIXct format
}
\description{
Extracts the time that dawn and dusk occur from a time series consisting of PAR values (light intensity).
Dawn is defined as the point where the par value goes from night -> day.
Dusk is defined as the point where the par value goes from day -> night.
This is useful for gating purposes for flow- cytometry.
Dawn and dusk tend to represent the extrema of flourences for phytoplankton.
}
\examples{
\dontrun{
par.data.csv <- system.file("extdata/par_data.csv", package="popcycle")
par.data <- read.csv(par.data.csv)
par.data$date <- as.POSIXct(par.data$date, format = "\%FT\%T", tz = "GMT")
dawn.dusk.times <- get.dawn.dusk.time(par.data, 10)
}
}
