% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.empty.evt.files}
\alias{get.empty.evt.files}
\title{Get names for EVT files which produced no OPP data.}
\usage{
get.empty.evt.files(db, evt.files)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{evt.files}{list of EVT file paths, e.g. get.evt.files(evt.dir).}
}
\value{
List of EVT files which produced no OPP data.
}
\description{
Get names for EVT files which produced no OPP data.
}
\examples{
\dontrun{
empty.evt.files <- get.empty.evt.files(db, evt.files)
}
}

