% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get.evt.by.file}
\alias{get.evt.by.file}
\title{Get EVT data frame by file.}
\usage{
get.evt.by.file(evt.dir, file.name, transform = FALSE)
}
\arguments{
\item{evt.dir}{EVT file directory.}

\item{file.name}{File name with julian day directory. Can be a single
file name or vector of file names.}

\item{transform}{Linearize EVT data. Default is FALSE.}
}
\value{
Data frame of EVT data for all files in file.name.
}
\description{
Get EVT data frame by file.
}
\examples{
\dontrun{
evt <- get.evt.by.file(evt.dir, "2014_185/2014-07-04T00-00-02+00-00")
}
}
