% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.evt.files.by.date}
\alias{get.evt.files.by.date}
\title{Get a list of EVT files by date range.}
\usage{
get.evt.files.by.date(db, evt.dir, start.date, end.date)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{evt.dir}{EVT file directory.}

\item{start.date}{Start date in format YYYY-MM-DD HH:MM.}

\item{end.date}{End date in format YYYY-MM-DD HH:MM.}
}
\value{
Vector of EVT files within date range.
}
\description{
Get a list of EVT files by date range.
}
\examples{
\dontrun{
evt.files <- get.evt.files.by.date(db, evt.dir, "2014-07-04 00:00", "2014-07-04 00:10")
}
}
