% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get.evt.files}
\alias{get.evt.files}
\title{Find EVT files with a recursive search down a directory tree.}
\usage{
get.evt.files(evt.dir)
}
\arguments{
\item{evt.dir}{Directory containing EVT files.}
}
\value{
Vector of EVT files with julian day directory.
}
\description{
Find EVT files with a recursive search down a directory tree.
}
\examples{
\dontrun{
evt.files <- get.evt.files(evt.dir)
}
}

