% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.filter.params.by.id}
\alias{get.filter.params.by.id}
\title{Get filter parameters by id.}
\usage{
get.filter.params.by.id(db, filter.id)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{filter.id}{ID for entry in filter table.}
}
\value{
Data frame of filter parameters matchign filter.id.
}
\description{
Get filter parameters by id.
}
\examples{
\dontrun{
filter.params <- get.filter.params.by.id(db, "d3afb1ea-ad20-46cf-866d-869300fe17f4")
}
}

