% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.filter.params.latest}
\alias{get.filter.params.latest}
\title{Get the latest filter parameters.}
\usage{
get.filter.params.latest(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Data frame of latest filter parameters.
}
\description{
Get the latest filter parameters.
}
\examples{
\dontrun{
filter.params <- get.filter.params.latest(db)
}
}

