% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.filter.table}
\alias{get.filter.table}
\title{Return a data frame for the filter table.}
\usage{
get.filter.table(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Data frame of filter table.
}
\description{
Return a data frame for the filter table.
}
\examples{
\dontrun{
filter.table <- get.filter.table(db)
}
}
