% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.gating.params.latest}
\alias{get.gating.params.latest}
\title{Get the latest gating parameters.}
\usage{
get.gating.params.latest(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Named list where list$gates.log is a recreation of original gating
  created by add.manual.classification(), add.auto.classification(), and
  save.gating.params()
}
\description{
Get the latest gating parameters.
}
\examples{
\dontrun{
gating.params <- get.gating.params.latest(db)
}
}
