% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.gating.table}
\alias{get.gating.table}
\title{Return a data frame for the gating table.}
\usage{
get.gating.table(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Data frame of gating table.
}
\description{
Return a data frame for the gating table.
}
\examples{
\dontrun{
gating.table <- get.gating.table(db)
}
}

