% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get.latest.evt}
\alias{get.latest.evt}
\title{Find the most recent EVT file.}
\usage{
get.latest.evt(evt.dir)
}
\arguments{
\item{evt.dir}{Directory containing EVT files.}
}
\value{
Most recent EVT file with julian day directory.
}
\description{
Find the most recent EVT file.
}
\examples{
\dontrun{
evt.file <- get.latest.evt(evt.dir)
}
}
