% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.meta.table}
\alias{get.meta.table}
\title{Get instrument serial number and cruise name}
\usage{
get.meta.table(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Data frame.
}
\description{
Get instrument serial number and cruise name
}
\examples{
\dontrun{
meta <- get.meta.table(db)
}
}
