% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R, R/files.R
\name{get.opp.by.date}
\alias{get.opp.by.date}
\title{Get OPP data frame by date and quantile}
\usage{
get.opp.by.date(
  db,
  opp.dir,
  start.date,
  end.date,
  outliers = TRUE,
  col_select = NULL
)

get.opp.by.date(
  db,
  opp.dir,
  start.date,
  end.date,
  outliers = TRUE,
  col_select = NULL
)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{opp.dir}{OPP file directory.}

\item{start.date}{Start date string, inclusive.}

\item{end.date}{End date string, inclusive.}

\item{outliers}{If TRUE, remove outliers. Otherwise don't perform outlier
filtering.}

\item{col_select}{col_select parameter passed to arrow::read_parquet as
tidy-selection. date and file_id are always added as the first columns.
By default all columns are returned.}

\item{quantile}{Filtering quantile for this file}

\item{channel}{Channels to keep in returned data frame. Can be a single name
or a vector. Choosing fewer channels can significantly speed up retrieval.}

\item{transform}{Linearize OPP data.}

\item{vct.dir}{VCT file directory. If not specified returned data frame will
not have a pop column.}

\item{pop}{If specified, the returned data frame will only contain entries
for this population.}
}
\value{
Data frame of OPP data for all files between start.date and end.date
  inclusive.

Data frame of filtered particles. If no data is found a data frame
  with zero rows will be returned.
}
\description{
Requires a populated sfl table.

Datetime string formats should be in the form "2020-07-31 10:00".
}
\examples{
\dontrun{
opp <- get.opp.by.date(db, opp.dir, "2014-07-04 00:00", "2014-07-04 00:10")
opp <- get.opp.by.date(db, opp.dir, "2014-07-04 00:00", "2014-07-04 00:10",
                       channel=c("fsc_small", "chl_small", "pe"),
                       transform=F, vct.dir=vct.dir)
}
\dontrun{
opp <- get.opp.by.date(db, opp.dir, "2020-07-31 10:00", "2020-07-31 12:00"
                       col_select=c("fsc_small"))
}
}
