% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get.opp.by.file}
\alias{get.opp.by.file}
\title{Get a tibble of filtered particles by file and quantile}
\usage{
get.opp.by.file(db, opp.dir, file.name, col_select = NULL)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{opp.dir}{OPP file directory.}

\item{file.name}{File name with julian day directory.}

\item{col_select}{col_select parameter passed to arrow::read_parquet as
tidy-selection. date and file_id are always added as the first columns.
By default all columns are returned.}
}
\value{
Tibble of filtered particles for one file. If no data is found a
  tibble with zero rows will be returned.
}
\description{
Get a tibble of filtered particles by file and quantile
}
\examples{
\dontrun{
opp <- get.opp.by.file(opp.dir, "2014_185/2014-07-04T00-00-02+00-00")
}
}
