% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.opp.dates}
\alias{get.opp.dates}
\title{Get ISO8601 datetime strings (e.g. 2018-07-13T22:24:40+00:00) for OPP files.}
\usage{
get.opp.dates(db, opp_files)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{opp_files}{Character vector of OPP file IDs.}
}
\value{
DataFrame with columns "file" and "date".
}
\description{
Get ISO8601 datetime strings (e.g. 2018-07-13T22:24:40+00:00) for OPP files.
}
\examples{
\dontrun{
opp_dates <- get.opp.dates(db, opp_files)
}
}
