% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.opp.files}
\alias{get.opp.files}
\title{Get OPP file names.}
\usage{
get.opp.files(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
List of OPP file names based on the latest filtering
  parameters or NULL if no filtering has been done.
}
\description{
Get OPP file names.
}
\examples{
\dontrun{
opp.files <- get.opp.files(db)
}
}
