% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{get.opp.gates}
\alias{get.opp.gates}
\title{Get per file gating IDs to use for regating based on previous gating.}
\usage{
get.opp.gates(db, opp_files, vct_table)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{opp_files}{List of OPP files to classify. Include julian day directory.}

\item{vct_table}{VCT table from past gating.}
}
\value{
DataFrame of "file", "opp_dates", "gating_id"
}
\description{
Based on the time ranges different gating parameters were applied during
a previous gating run (as captured by vct_table), return a DataFrame with
OPP file to gating ID matches.
}
\examples{
\dontrun{
get.opp.gates(db, opp_files, vct_table)
}
}
