% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.opp.stats.table}
\alias{get.opp.stats.table}
\title{Return a data frame for the opp table.}
\usage{
get.opp.stats.table(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Data frame of opp table.
}
\description{
Return a data frame for the opp table.
}
\examples{
\dontrun{
opp.stats.table <- get.opp.table(db)
}
}

