% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.outlier.table}
\alias{get.outlier.table}
\title{Get list of outliers.}
\usage{
get.outlier.table(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Data frame.
}
\description{
Get list of outliers.
}
\examples{
\dontrun{
outlier.table <- get.outlier.table(db)
}
}
