% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.poly.log.by.gating.id.pop}
\alias{get.poly.log.by.gating.id.pop}
\title{Construct a gating polygon list for gating.id pop combo}
\usage{
get.poly.log.by.gating.id.pop(db, gating.id, popname)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{gating.id}{Foreign key to gating table.}

\item{popname}{Population name}
}
\value{
List of population gating polygon coordinates.
}
\description{
Construct a gating polygon list for gating.id pop combo
}
\examples{
\dontrun{
poly.log <- get.poly.log.by.gating.id.pop(db, "d3afb1ea-ad20-46cf-866d-869300fe17f4")
}
}

