% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.poly.log.by.gating.id}
\alias{get.poly.log.by.gating.id}
\title{Construct a gating polygon named list for gating.id}
\usage{
get.poly.log.by.gating.id(db, gating.id, pop.order)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{gating.id}{Foreign key to gating table.}

\item{pop.order}{Comma-separated text specifying population order. This is
the pop_order field in the gating table.}
}
\value{
List of population gating polygon coordinates.
}
\description{
Construct a gating polygon named list for gating.id
}
\examples{
\dontrun{
poly.log <- get.poly.log.by.gating.id(db, "d3afb1ea-ad20-46cf-866d-869300fe17f4",
                                      "beads,prochloro,synecho")
}
}

