% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.poly.table}
\alias{get.poly.table}
\title{Return a data frame for the poly table.}
\usage{
get.poly.table(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Data frame of poly table.
}
\description{
Return a data frame for the poly table.
}
\examples{
\dontrun{
poly.table <- get.poly.table(db)
}
}
