% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.sfl.by.date}
\alias{get.sfl.by.date}
\title{Get SFL rows >= start.date and <= end.date.}
\usage{
get.sfl.by.date(db, start.date, end.date)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{start.date}{Start date in format YYYY-MM-DD HH:MM.}

\item{end.date}{End date in format YYYY-MM-DD HH:MM.}
}
\value{
Data frame of sfl table entries between start.date and end.date
  inclusive.
}
\description{
Get SFL rows >= start.date and <= end.date.
}
\examples{
\dontrun{
sfl <- get.sfl.by.date(db, "2014-07-04 00:00", "2014-07-04 00:10")
}
}
