% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.sfl.table}
\alias{get.sfl.table}
\title{Return a data frame for the sfl table.}
\usage{
get.sfl.table(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Data frame of sfl table.
}
\description{
Return a data frame for the sfl table.
}
\examples{
\dontrun{
sfl.table <- get.sfl.table(db)
}
}

