% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get.vct.by.date}
\alias{get.vct.by.date}
\title{Get data frame of per particle population classifications selecting by date.}
\usage{
get.vct.by.date(
  db,
  vct.dir,
  start.date,
  end.date,
  outliers = TRUE,
  col_select = NULL
)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{vct.dir}{VCT file directory.}

\item{start.date}{Start date string, inclusive.}

\item{end.date}{End date string, inclusive.}

\item{outliers}{If TRUE, remove outliers. Otherwise don't perform outlier
filtering.}

\item{col_select}{col_select parameter passed to arrow::read_parquet as
tidy-selection. date and file_id are always added as the first columns.
By default all columns are returned.}
}
\value{
Data frame of per particle population classifications. If no data is
  found a data frame with zero rows will be returned.
}
\description{
Datetime string formats should be in the form "2020-07-31 10:00".
}
\examples{
\dontrun{
vct <- get.vct.by.date(db, vct.dir, "2020-07-31 10:00", "2020-07-31 12:00"
                       col_select=c(ends_with("q50"), starts_with("diam_mid"))
}
}
