% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get.vct.by.file}
\alias{get.vct.by.file}
\title{Get a tibble of per particle population classifications by file and quantile}
\usage{
get.vct.by.file(db, vct.dir, file.name, col_select = NULL)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{vct.dir}{VCT file directory.}

\item{file.name}{File name with julian day directory.}

\item{col_select}{col_select parameter passed to arrow::read_parquet as
tidy-selection. date and file_id are always added as the first columns.
By default all columns are returned.}
}
\value{
Tibble of per particle population classifications. If no data is
  found a tibble with zero rows will be returned.
}
\description{
Get a tibble of per particle population classifications by file and quantile
}
\examples{
\dontrun{
vct <- get.vct.by.file(vct.dir, "2014_185/2014-07-04T00-00-02+00-00")
}
}
