% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.vct.files}
\alias{get.vct.files}
\title{Get VCT file names.}
\usage{
get.vct.files(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
List of distinct VCT file names.
}
\description{
Get VCT file names.
}
\examples{
\dontrun{
vct.files <- get.vct.files(db)
}
}

