% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.vct.stats.by.date}
\alias{get.vct.stats.by.date}
\title{Get VCT aggregated per population statistics by date.}
\usage{
get.vct.stats.by.date(db, start.date, end.date)
}
\arguments{
\item{db}{SQLite3 data file path.}

\item{start.date}{Start date in format YYYY-MM-DD HH:MM.}

\item{end.date}{End date in format YYYY-MM-DD HH:MM.}
}
\value{
Data frame of VCT aggregate statistics for all files between
  start.date and end.date inclusive.
}
\description{
Requires a populated sfl table.
}
\examples{
\dontrun{
vct.stats <- get.vct.stats.by.date(db, "2014-07-04 00:00", "2014-07-04 00:10")
}
}
