% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.vct.stats.by.file}
\alias{get.vct.stats.by.file}
\title{Get VCT aggregated per population statistics by file.}
\usage{
get.vct.stats.by.file(db, file.name)
}
\arguments{
\item{db}{SQLite3 data file path.}

\item{file.name}{File name with julian day directory.}
}
\value{
Data frame of VCT aggregate statistics for one file.
}
\description{
Requires a populated sfl table.
}
\examples{
\dontrun{
vct.stats <- get.vct.stats.by.file(db, "2014_185/2014-07-04T00-00-02+00-00")
}
}
