% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.vct.table}
\alias{get.vct.table}
\title{Return a data frame for the vct table.}
\usage{
get.vct.table(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Data frame of vct table.
}
\description{
Return a data frame for the vct table.
}
\examples{
\dontrun{
vct.table <- get.vct.table(db)
}
}

