% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get_evt_by_file}
\alias{get_evt_by_file}
\title{Get EVT data frame by file.}
\usage{
get_evt_by_file(evt_dir, file_id, transform = FALSE)
}
\arguments{
\item{file_id}{File ID with julian day directory. Can be a single
file name or vector of file names.}

\item{transform}{Linearize EVT data. Default is FALSE.}

\item{evt.dir}{EVT file directory.}
}
\value{
Data frame of EVT data for all files in file.name.
}
\description{
Get EVT data frame by file.
}
