% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get_evt_files}
\alias{get_evt_files}
\title{Find EVT files with a recursive search down a directory tree.}
\usage{
get_evt_files(evt_dir, db = NULL)
}
\arguments{
\item{evt_dir}{Directory containing EVT files.}

\item{db}{SQLite3 popcycle database file. If present, inner join to sfl table
and add a date column.}
}
\value{
Tibble of EVT file paths, file IDs, and optionally dates: ("path",
  "file_id", "date").
}
\description{
Find EVT files with a recursive search down a directory tree.
}
