% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_filter_params_by_id}
\alias{get_filter_params_by_id}
\title{Get filter parameters by id.}
\usage{
get_filter_params_by_id(db, filter_id)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{filter_id}{ID for entry in filter table.}
}
\value{
Data frame of filter parameters matchign filter_id.
}
\description{
Get filter parameters by id.
}
