% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_filter_plan_table}
\alias{get_filter_plan_table}
\title{Return a tibble for the filter_plan table.}
\usage{
get_filter_plan_table(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Tibble of filter plan table
}
\description{
Return a tibble for the filter_plan table.
}
