% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_gating_params_by_id}
\alias{get_gating_params_by_id}
\title{Get gating parameters by ID.}
\usage{
get_gating_params_by_id(db, gating_id)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{gating_id}{ID in gating table and poly table.}
}
\value{
Named list where list$gates.log a recreation of original gating
  created by add_manual_classification(), add_auto_classification(), and
  save_gating_params()
}
\description{
Get gating parameters by ID.
}
