% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_gating_plan_table}
\alias{get_gating_plan_table}
\title{Return a tibble for the gating_plan table.}
\usage{
get_gating_plan_table(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Tibble of gating plan table.
}
\description{
Return a tibble for the gating_plan table.
}
