% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_inst}
\alias{get_inst}
\title{Get instrument serial number}
\usage{
get_inst(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
One serial number
}
\description{
Get instrument serial number
}
