% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_metadata_table}
\alias{get_metadata_table}
\title{Return a tibble of the metadata table of cruise and instrument serial.}
\usage{
get_metadata_table(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Tibble for metadata table
}
\description{
Return a tibble of the metadata table of cruise and instrument serial.
}
