% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_opp2_table}
\alias{get_opp2_table}
\title{Return a tibble for the opp2 table.}
\usage{
get_opp2_table(
  db,
  sfl_join = TRUE,
  all_sfl_columns = FALSE,
  outlier_join = TRUE,
  file_flag_filter = FALSE
)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{sfl_join}{Join to sfl table by file ID, adding a date column and
removing opp2 entries with no corresponding SFL entry.}

\item{all_sfl_columns}{If joining to sfl, include all sfl columns.}

\item{outlier_join}{Left join to outlier table by file ID, adding an outlier
flag column. This function will not perform any filtering by outlier flag.}

\item{file_flag_filter}{Filter by file_flag column. If TRUE, only keep files with
file_flag == 0.}
}
\value{
Tibble of opp2 table.
}
\description{
If the opp2 table does not exist in the database this function will return an
empty tibble.
}
