% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get_opp_by_date}
\alias{get_opp_by_date}
\title{Get tibble of filtered particles selecting by date.}
\usage{
get_opp_by_date(
  db,
  opp_dir,
  start_date,
  end_date,
  outliers = TRUE,
  col_select = NULL
)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{opp_dir}{OPP file directory.}

\item{start_date}{Start date, inclusive.}

\item{end_date}{End date, inclusive.}

\item{outliers}{If TRUE, remove outliers. Otherwise don't perform outlier
filtering.}

\item{col_select}{col_select parameter passed to arrow::read_parquet as
a character vector. date and file_id are always added as the first columns.
By default all columns are returned.}
}
\value{
Tibble of filtered particles. If no data is found an empty tibble
  will be returned.
}
\description{
Datetime string formats should be in the form "2020-07-31 10:00".
}
