% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_opp_table}
\alias{get_opp_table}
\title{Return a tibble for the opp table.}
\usage{
get_opp_table(
  db,
  sfl_join = TRUE,
  all_sfl_columns = FALSE,
  outlier_join = TRUE,
  particles_in_all_quantiles = TRUE,
  file_flag_filter = TRUE
)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{sfl_join}{Join to SFL table by file ID, adding a date column and
removing OPP entries with no corresponding SFL entry.}

\item{all_sfl_columns}{If joining to SFL, include all SFL columns.}

\item{outlier_join}{Left join to outlier table by file ID, adding an outlier
flag column. This function will not perform any filtering by outlier flag.}

\item{particles_in_all_quantiles}{Only keep files that have OPP particles in
all quantiles.}
}
\value{
Tibble of opp table.
}
\description{
Return a tibble for the opp table.
}
