% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_outlier_table}
\alias{get_outlier_table}
\title{Get tibble of outliers.}
\usage{
get_outlier_table(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Tibble of outlier table.
}
\description{
Get tibble of outliers.
}
