% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_poly_log_by_gating_id_pop}
\alias{get_poly_log_by_gating_id_pop}
\title{Construct a gating polygon list for gating_id pop combo}
\usage{
get_poly_log_by_gating_id_pop(db, gating_id, popname)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{gating_id}{Foreign key to gating table.}

\item{popname}{Population name}
}
\value{
List of population gating polygon coordinates.
}
\description{
Construct a gating polygon list for gating_id pop combo
}
