% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_poly_table}
\alias{get_poly_table}
\title{Return a data frame for the poly table.}
\usage{
get_poly_table(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Data frame of poly table.
}
\description{
Return a data frame for the poly table.
}
