% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get_processed_particles}
\alias{get_processed_particles}
\title{Get a tibble of OPP or VCT particles}
\usage{
get_processed_particles(
  db,
  data_dir,
  data_type,
  file_ids = NULL,
  start_date = NULL,
  end_date = NULL,
  outliers = TRUE,
  col_select = NULL
)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{data_dir}{Directory with OPP or VCT parquet files.}

\item{data_type}{"opp" or "vct".}

\item{file_ids}{Vector of file IDs with julian day directory.}

\item{start_date}{Start date, inclusive.}

\item{end_date}{End date, inclusive.}

\item{outliers}{If TRUE, remove outliers. Otherwise don't perform outlier
filtering.}

\item{col_select}{col_select parameter passed to arrow::read_parquet as
a charactr vector. date and file_id are always added to to this vector. By
default all columns are returned.}
}
\value{
Tibble of OPP or VCT particles. If no data is found a tibble with
  zero rows will be returned.
}
\description{
Get a tibble of OPP or VCT particles
}
