% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_raw_stat_table}
\alias{get_raw_stat_table}
\title{Get aggregate statistics data frame joining sfl, opp, and vct table entries.}
\usage{
get_raw_stat_table(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Data frame of aggregate statistics.
}
\description{
Get aggregate statistics data frame joining sfl, opp, and vct table entries.
}
