% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_sfl_table}
\alias{get_sfl_table}
\title{Return a tibble for the sfl table.}
\usage{
get_sfl_table(db, outlier_join = TRUE)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{outlier_join}{Left join to outlier table by file ID, adding an outlier
flag column. This function will not perform any filtering by outlier flag.}
}
\value{
Tibble of sfl table.
}
\description{
Return a tibble for the sfl table.
}
