% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_stat_table}
\alias{get_stat_table}
\title{Get aggregate statistics data frame along with estimates of cell abundance.}
\usage{
get_stat_table(db, inst = NULL)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{inst}{Instrument serial. If not provided will attempt to read from db.}
}
\value{
Data frame of aggregate statistics.
}
\description{
Get aggregate statistics data frame along with estimates of cell abundance.
}
