% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get_vct_by_file}
\alias{get_vct_by_file}
\title{Get tibble of classified particles selecting by file IDs.}
\usage{
get_vct_by_file(db, vct_dir, file_ids, outliers = TRUE, col_select = NULL)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{vct_dir}{VCT file directory.}

\item{file_ids}{Vector of file IDs with julian day directory.}

\item{outliers}{If TRUE, remove outliers. Otherwise don't perform outlier
filtering.}

\item{col_select}{col_select parameter passed to arrow::read_parquet as
a character vector. date and file_id are always added as the first columns.
By default all columns are returned.}
}
\value{
Tibble of classified particles. If no data is found an empty tibble
  will be returned.
}
\description{
Get tibble of classified particles selecting by file IDs.
}
