% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{get_vct_quantile_range}
\alias{get_vct_quantile_range}
\title{Find quantiles for VCT data}
\usage{
get_vct_quantile_range(
  vct_files,
  col,
  filtering_quantile,
  quantile_probs = c(0.01, 0.99),
  pop = NULL,
  ignore_dates = NULL
)
}
\arguments{
\item{vct_files}{VCT file paths.}

\item{filtering_quantile}{OPP Filtering quantile.}

\item{quantile_probs}{Numeric vector of probabilities with values in [0,1].}

\item{ignore_dates}{Don't process VCT data with these dates.}

\item{data_cols}{Character vector of VCT column to use as size metric:
fsc_small, chl_small, pe, Qc_[lwr,mid,upr], or diam_[lwr,mid,upr].
For Qc and diam the filtering quantile will be added to the column name
automatically.}

\item{pops}{Single population label to filter for.}
}
\value{
Result of quantile()
}
\description{
Find quantiles for VCT data
}
