% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{grid_bins_labels}
\alias{grid_bins_labels}
\title{Construct string labels for break points created by create_grid_bins()}
\usage{
grid_bins_labels(grid_bins)
}
\arguments{
\item{grid_bins}{Named list of breakpoints created by create_grid_bins().}
}
\value{
Named list identical to grid with numeric break points converted to strings.
}
\description{
Construct string labels for break points created by create_grid_bins()
}
