% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{grid_labels}
\alias{grid_labels}
\title{Construct string labels for break points created by create_grid()}
\usage{
grid_labels(grid)
}
\arguments{
\item{grid}{Named list of breakpoints created by create_grid().}
}
\value{
Named list identical to grid with numeric break points converted to strings.
}
\description{
Construct string labels for break points created by create_grid()
}
