% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{group_gridded_by_time}
\alias{group_gridded_by_time}
\title{Group gridded particle data at a lower time resolution}
\usage{
group_gridded_by_time(gridded, time_expr = "1 hours", use_data.table = TRUE)
}
\arguments{
\item{gridded}{Gridded data created by grid().}

\item{time_expr}{Time expression passed to lubridate::floor_date to
lower time resolution.}

\item{use_data.table}{Use data.table for performance speedup, otherwise use dplyr.}
}
\value{
A tibble of gridded with reduced time resolution.
}
\description{
Group gridded particle data at a lower time resolution
}
