% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{identify_noise}
\alias{identify_noise}
\title{Return logical index vector for noise particle rows}
\usage{
identify_noise(evt)
}
\arguments{
\item{evt}{EVT data frame}
}
\value{
Logical index vector for noise particle rows
}
\description{
Return logical index vector for noise particle rows
}
