% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{identify_saturated}
\alias{identify_saturated}
\title{Return logical index vector for saturated particle rows}
\usage{
identify_saturated(evt)
}
\arguments{
\item{evt}{EVT data frame}
}
\value{
Logical index vector for saturated particle rows
}
\description{
Return logical index vector for saturated particle rows
}
