% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{inflection.point}
\alias{inflection.point}
\title{Plot helpful cytograms for estimating the D1, D2 and FSC coordinates of the inflection point (corresponds to location of 1µm beads).}
\usage{
inflection.point(DF)
}
\arguments{
\item{dataframe}{containing EVT data.}
}
\value{
D1, D2 and fsc values of presumed 1 µm beads
}
\description{
Plot helpful cytograms for estimating the D1, D2 and FSC coordinates of the inflection point (corresponds to location of 1µm beads).
}
