% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{inflection_point}
\alias{inflection_point}
\title{Estimate coordinates of the inflection point (location of 1µm beads).}
\usage{
inflection_point(df)
}
\arguments{
\item{df}{Data frame containing log (untransformed) EVT data}
}
\value{
25th, 50th, 75th quantiles for D1, D2 and fsc values of presumed 1µm
  beads, labeled as quantiles "2.5", "50", "97.5".
}
\description{
Estimate coordinates of the inflection point (location of 1µm beads).
}
