% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\name{is_transformed}
\alias{is_transformed}
\title{Check if an EVT data frame has been transformed.}
\usage{
is_transformed(df)
}
\arguments{
\item{df}{EVT data frame}
}
\value{
Boolean indicating if df is transformed
}
\description{
Check if an EVT data frame has been transformed.
}
