% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{make.popcycle.db}
\alias{make.popcycle.db}
\title{Create a new, empty sqlite3 popcycle database.}
\usage{
make.popcycle.db(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
None
}
\description{
If a database already exists missing tables and indexes will be added.
}
\examples{
\dontrun{
make.popcycle.db(db)
}
}

