% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{make_popcycle_db}
\alias{make_popcycle_db}
\title{Create a new, empty sqlite3 popcycle database.}
\usage{
make_popcycle_db(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
None
}
\description{
If a database already exists missing tables and indexes will be added.
}
