% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{manual.classify}
\alias{manual.classify}
\title{Classify particles based on manually defined population gates.}
\usage{
manual.classify(opp, params, popname)
}
\arguments{
\item{opp}{OPP data frame.}

\item{params}{Named list of gating parameters. Must contain a params$poly
entry with polygon definitions.}
}
\value{
List of per particle classifications.
}
\description{
Classify particles based on manually defined population gates.
}
\examples{
\dontrun{
vct <- manual.classify(opp, gates.log, "beads")
}
}

