% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{ManualGating}
\alias{ManualGating}
\title{Classify particles based on manually defined population gates.}
\usage{
ManualGating(opp, poly.log)
}
\arguments{
\item{opp}{OPP data frame.}

\item{poly.log}{Named list of gating polygon definitions. Particles will be
classified by population in the order they appear in poly.log.}
}
\value{
List of per particle classifications.
}
\description{
Classify particles based on manually defined population gates.
}
\examples{
\dontrun{
vct <- ManualGating(opp, poly.log)
}
}

