% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{merge.stat}
\alias{merge.stat}
\title{Calculate mean  + sd of aggregate statistics data from the 3 OPP data sets (2.5, 50 and 97.5 qunatile).}
\usage{
\method{merge}{stat}(stat)
}
\arguments{
\item{stat.}{Table generated by the function carbon_conversion.}
}
\value{
Data frame of aggregate statistics.
}
\description{
Calculate mean  + sd of aggregate statistics data from the 3 OPP data sets (2.5, 50 and 97.5 qunatile).
}
\examples{
\dontrun{
DF <- merge.stat(stat)
}
}
