% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{merge_and_reanalyze}
\alias{merge_and_reanalyze}
\title{Merge databases after refiltering and reanalyze OPP.}
\usage{
merge_and_reanalyze(dir_old, dir_new, mie_table = NULL)
}
\arguments{
\item{dir_old}{Directory with old popcycle databases. Database files will be
searched for recursively. Only the database files are used. Other results
files are ignored.}

\item{dir_new}{Directory with new filtered results. Database files will be
searched for recursively. For each database file it is expected that there
will be corresponding OPP directories in the same location, with a naming
scheme of <cruise>_opp/. <cruise> should match the cruise stored in the
metadata table.}

\item{mie_table}{Mie theory lookup table to use in place of installed table.}
}
\value{
None
}
\description{
This functions automates reanalysis that occurs downstream from raw particle
filtration, e.g. gating, diameter, carbon. It will clear the gating, poly,
and metadata tables in databases in dir_new and copy the same tables from
corresponding databases found in dir_old. It will also copy outlier flags
from old databases to new for files that exist in both databases. After the
new databases have been updated, OPP particles will be reclassified and
diameters and carbon quotas will be recalculated.
}
\examples{
\dontrun{
merge_and_reanalyze(dir_old, dir_new)
}
}
