% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{merge_old_new_vct}
\alias{merge_old_new_vct}
\title{Merge two VCT dataframes by adding rows for novel file_id in old into new.}
\usage{
merge_old_new_vct(old_vct_df, new_vct_df)
}
\arguments{
\item{old_vct_df}{VCT to pull novel file_id entries from.}

\item{new_vct_df}{VCT into which novel file_id entries will be added.}
}
\value{
new_vct_df with novel file_id entries from old_vct_df.
}
\description{
Merge two VCT dataframes by adding rows for novel file_id in old into new.
}
