% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{migrate_tables}
\alias{migrate_tables}
\title{Migrate filtering, gating, and outlier related tables to a new post-filter database.}
\usage{
migrate_tables(old_db, new_db)
}
\arguments{
\item{db_from}{Database to copy outlier table from.}

\item{db_to}{Database to copy outlier table to.}
}
\description{
Tables migrated are filter, filter_plan, gating_plan, gating, poly, and outlier.
}
