% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{normalization}
\alias{normalization}
\title{Normalization by beads used as internal standard}
\usage{
normalization(stat, spar = 0.1)
}
\arguments{
\item{stat}{Stat table generated by the function get.stat.table().}

\item{spar}{smooothing parameter, the higher the more smoothing is applied.}
}
\value{
A dataframe with normalized D1, D2, fsc_small, chl_small, pe and fsc_perp values
}
\description{
Normalization by beads used as internal standard
}
