% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{opp_quantile_inner_join_clause}
\alias{opp_quantile_inner_join_clause}
\title{Add an INNER JOIN to only select OPP files with data in all quantiles.}
\usage{
opp_quantile_inner_join_clause()
}
\value{
SQL INNER JOIN string
}
\description{
Return a SQL string with an INNER JOIN to a subquery selecting for OPP files
that contain data in all quantiles. The benefit of making this a subquery is
the GROUP BY doesn't affect the rest of the SQL query this JOIN is embedded
into.
}
\examples{
\dontrun{
sql <- opp_quantile_inner_join_clause()
}
}
