% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{parse_file_dates}
\alias{parse_file_dates}
\title{Convert timestamp component of SeaFlow filenames to POSIXct.}
\usage{
parse_file_dates(file_timestamps)
}
\arguments{
\item{file_timestamps}{Character vector of timestamp components of SeaFlow
filenames matching the format YYYY-MM-DDTHH-MM-SS+HH-MM, for example
"2020-07-23T14-04-05+00-00". This corresponds to the RFC3339 formatted
timestamp "2020-07-23T14:04:05+00:00".}
}
\value{
POSIXct vector
}
\description{
Convert timestamp component of SeaFlow filenames to POSIXct.
}
\examples{
\dontrun{
dates <- parse_file_dates(
  c("2020-07-23T14:04:05+00:00", "2020-07-25T14:04:05+00:00")
)
}
}
