% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan_from_table}
\alias{plan_from_table}
\title{Create a tibble of start dates for gating / filter IDs based vct / opp table entries.}
\usage{
plan_from_table(tbl)
}
\arguments{
\item{tbl}{VCT or OPP table as data.frame or tibble}
}
\value{
tibble with character columns start_date and gating_id
}
\description{
Create a tibble of start dates for gating / filter IDs based vct / opp table entries.
}
