% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.cytogram}
\alias{plot.cytogram}
\title{Plot cytogram.}
\usage{
plot.cytogram(opp, para.x = "fsc_small", para.y = "chl_small", ...)
}
\arguments{
\item{opp}{OPP data frame.}

\item{para.x}{Channel to use as x axis.}

\item{para.y}{Channel to use as y axis.}

\item{...}{Additional parameters for densCols()}
}
\value{
None
}
\description{
Plot cytogram.
}

