% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.evt.cytogram.by.file}
\alias{plot.evt.cytogram.by.file}
\title{Plot cytogram for one EVT file.}
\usage{
plot.evt.cytogram.by.file(evt.dir, file.name, para.x = "fsc_small",
  para.y = "chl_small", ...)
}
\arguments{
\item{evt.dir}{EVT file directory.}

\item{file.name}{File name with julian day directory.}

\item{para.x}{Channel to use as x axis.}

\item{para.y}{Channel to use as y axis.}

\item{...}{Additional parameters for densCols()}
}
\value{
None
}
\description{
Plot cytogram for one EVT file.
}

