% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{plot.filter.cytogram.by.file}
\alias{plot.filter.cytogram.by.file}
\title{Plot helpful cytograms for exploring filtering parameters for one file.}
\usage{
plot.filter.cytogram.by.file(evt.dir, file.name, filter.params)
}
\arguments{
\item{evt.dir}{EVT file directory.}

\item{file.name}{File name with julian day directory.}

\item{filter.params}{Filtering parameters in a one row data frame or named
list. Columns should include width, notch.small.D1, notch.small.D2,
notch.large.D1, notch.large.D2, offset.small.D1, offset.small.D2,
offset.large.D1, offset.large.D2.}
}
\value{
None
}
\description{
Plot helpful cytograms for exploring filtering parameters for one file.
}
