% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.gating.cytogram.by.file}
\alias{plot.gating.cytogram.by.file}
\title{Plot cytogram with gates for one file.}
\usage{
plot.gating.cytogram.by.file(opp.dir, file.name, poly.log = NULL,
  para.x = "fsc_small", para.y = "chl_small")
}
\arguments{
\item{opp.dir}{OPP file directory.}

\item{file.name}{File name with julian day directory.}

\item{poly.log}{Gating polygon(s) to draw (optional).}

\item{para.x}{Channel to use as x axis.}

\item{para.y}{Channel to use as y axis.}
}
\value{
None
}
\description{
Plot cytogram with gates for one file.
}
