% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.gating.cytogram}
\alias{plot.gating.cytogram}
\title{Plot cytogram with gates.}
\usage{
plot.gating.cytogram(opp, poly.log = NULL, para.x = "fsc_small",
  para.y = "chl_small")
}
\arguments{
\item{opp}{OPP data frame.}

\item{poly.log}{Gating polygon(s) to draw (optional)}

\item{para.x}{Channel to use as x axis.}

\item{para.y}{Channel to use as y axis.}
}
\value{
None
}
\description{
Plot cytogram with gates.
}
