% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.histogram}
\alias{plot.histogram}
\title{Plot histogram.}
\usage{
plot.histogram(opp, para.x='fsc_small', transform=T)
}
\arguments{
\item{evtopp}{EVT or OPP data frame.}

\item{para.x}{Channel to use as x axis.}

\item{transform}{Log transformation of the parameter'}

\item{position}{Position adjustment, either as a string ('stack' or 'identity'), or the result of a call to a position adjustment function.}

\item{free}{Should the y-scale be free (TRUE) or fixed (FIXED)}
}
\value{
None
}
\description{
Plot histogram.
}
