% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.map}
\alias{plot.map}
\title{Plot cell abundances of a population on a map.}
\usage{
plot.map(stat, param, ...)
}
\arguments{
\item{stat}{Stat table from get.stat.table function}

\item{param}{Parameter to display}

\item{transform}{Log transformation of the parameter'}
}
\value{
None
}
\description{
Plot cell abundances of a population on a map.
}
