% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.vct.cytogram.by.file}
\alias{plot.vct.cytogram.by.file}
\title{Plot cytogram with particles colored by population for one file.}
\usage{
plot.vct.cytogram.by.file(opp.dir, vct.dir, file.name,
  para.x = "fsc_small", para.y = "chl_small", ...)
}
\arguments{
\item{opp.dir}{OPP file directory.}

\item{vct.dir}{VCT file directory.}

\item{file.name}{File name with julian day directory.}

\item{para.x}{Channel to use as x axis.}

\item{para.y}{Channel to use as y axis.}

\item{...}{Additional parameters for densCols()}
}
\value{
None
}
\description{
Plot cytogram with particles colored by population for one file.
}
