% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_cyt}
\alias{plot_cyt}
\title{Plot EVT or OPP cytogram with only builtin R graphics.}
\usage{
plot_cyt(evtopp, para.x = "fsc_small", para.y = "chl_small", ...)
}
\arguments{
\item{evtopp}{EVT or OPP data frame.}

\item{para.x}{Channel to use as x axis.}

\item{para.y}{Channel to use as y axis.}

\item{...}{Additional parameters for plot()}
}
\value{
None
}
\description{
Plot EVT or OPP cytogram with only builtin R graphics.
}
