% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_cytogram}
\alias{plot_cytogram}
\title{Plot EVT or OPP cytogram.}
\usage{
plot_cytogram(evtopp, para.x = 'fsc_small', para.y = 'chl_small', ...)
}
\arguments{
\item{evtopp}{EVT or OPP data frame.}

\item{para.x}{Channel to use as x axis.}

\item{para.y}{Channel to use as y axis.}

\item{bins}{Number of bin (hex) to display.}

\item{transform}{Log transformation for both x- and y-axis'}
}
\value{
None
}
\description{
Plot EVT or OPP cytogram.
}
