% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_filter_cytogram}
\alias{plot_filter_cytogram}
\title{Plot cytograms for exploring filtering parameters.}
\usage{
plot_filter_cytogram(evt, filter.params)
}
\arguments{
\item{evt}{EVT data frame.}

\item{filter.params}{Filtering parameters in a one row data frame or named
list. Columns should include width, notch.small.D1, notch.small.D2,
notch.large.D1, notch.large.D2, offset.small.D1, offset.small.D2,
offset.large.D1, offset.large.D2.}
}
\value{
None
}
\description{
Plot cytograms for exploring filtering parameters.
}
