% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Plot histogram.}
\usage{
plot_histogram(opp, para.x="fsc_small", transform=T)
}
\arguments{
\item{evtopp}{EVT or OPP data frame.}

\item{para.x}{Channel to use as x axis.}

\item{binwidth}{The width of the bins. Can be specified as a numeric value, or a function that calculates width from x. T
The default is to use bins bins that cover the range of the data. You should always override this value, exploring multiple widths to find the best to illustrate the stories in your data.}

\item{transform}{Log transformation of the parameter"}

\item{position}{Position adjustment, either as a string ("stack" or "identity"), or the result of a call to a position adjustment function.}

\item{free}{Should the y-scale be free (TRUE) or fixed (FIXED)}
}
\value{
None
}
\description{
Plot histogram.
}
