% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_map}
\alias{plot_map}
\title{Plot population distribution on a map.}
\usage{
plot_map(stat, param)
}
\arguments{
\item{stat}{Stat table from get.stat.table function}

\item{param}{Parameter to display}

\item{transform}{Log transformation of the parameter"}
}
\value{
None
}
\description{
Plot population distribution on a map.
}
