% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{POSIXct.to.db.date}
\alias{POSIXct.to.db.date}
\title{Convert a POSIXct date objectformat suitable for popcycle db date comparison.}
\usage{
POSIXct.to.db.date(date.ct)
}
\arguments{
\item{date.ct}{POSIXct date object.}
}
\value{
Date text as YYYY-MM-DDTHH:MM:00.
}
\description{
Convert a POSIXct date objectformat suitable for popcycle db date comparison.
}
\examples{
\dontrun{
db.date <- POSIXct.to.db.date(date.ct)
}
}
