% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{PSD.calibration}
\alias{PSD.calibration}
\title{Calibrate abundance based on Influx data}
\usage{
PSD.calibration(PSD, cruisename, calib = NULL)
}
\arguments{
\item{PSD}{Particle size disitribution created by create_PSD().}

\item{cruisename}{Name of the cruise (returned from get.mata.table()).}

\item{calib}{Optional calibration calibration  dataframe of regression values (cruise=cruise.name, pop=prochloro, a=slope, b=intercept)}
}
\value{
the stat table with corrected abundancce if available
}
\description{
Calibrate abundance based on Influx data
}
