% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{psd_abundance}
\alias{psd_abundance}
\title{Add metadata to a particle size distribution and convert raw counts to abundance}
\usage{
psd_abundance(db, distribution, quantile)
}
\arguments{
\item{db}{Popcycle database file path}

\item{distribution}{Distribution dataframe created by particle_counts_one_file()}

\item{quantile}{OPP Filtering quantile.}
}
\value{
distribution with counts converted to abundance, opp and sfl metadata merged.
}
\description{
Add metadata to a particle size distribution and convert raw counts to abundance
}
\examples{
\dontrun{
psd <- psd_abundance(db, psd, 2.5)
}
}
