% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_calib_csv}
\alias{read_calib_csv}
\title{Read Abundance calibration file}
\usage{
read_calib_csv(path = NULL)
}
\value{
A dataframe of regression values (cruise=cruise.name, pop=prochloro, a=slope, b=intercept)
}
\description{
Read Abundance calibration file
}
