% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_mie_csv}
\alias{read_mie_csv}
\title{Read Mie theory calibration file}
\usage{
read_mie_csv(path = NULL)
}
\value{
A dataframe of Mie theory conversion values
}
\description{
Read Mie theory calibration file
}
