% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_par_csv}
\alias{read_par_csv}
\title{Read table of PAR calibration values}
\usage{
read_par_csv(path = NULL)
}
\value{
A dataframe with columns for cruise and PAR calibration factor
}
\description{
Read table of PAR calibration values
}
