% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_parquet_one_quantile}
\alias{read_parquet_one_quantile}
\title{Read an OPP/VCT parquet file, only grabbing columns needed to analyze one quantile.}
\usage{
read_parquet_one_quantile(filepath, quantile, cols, refracs = NULL)
}
\arguments{
\item{filepath}{Path of parquet file}

\item{quantile}{Quantile to select, e.g. 2.5}

\item{cols}{Columns to select. If a column has a refractive index suffix and/or
a quantile suffix in the parquet file column name, don't include it here.
For example, use "pop", not "pop_q2.5". Use "diam", not "diam_lwr_q2.5". The
appropriate quantile boolean column will be added automatically if not present.
If not a character vector, all columns will be returned.}

\item{refracs}{Dataframe of refractive indices to use for each population. Column names
should be population names used in classification, values should be mid, lwr, upr.
If refracs is supplied but "pop" is not in the cols parameter, it will be added automatically.}
}
\description{
Read an OPP/VCT parquet file, only grabbing columns needed to analyze one quantile.
}
