% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_reference_filter_params}
\alias{read_reference_filter_params}
\title{Read table of reference filter parameters for a single instrument}
\usage{
read_reference_filter_params(inst = NULL)
}
\arguments{
\item{inst}{Instrument serial}
}
\value{
A dataframe of filter parameters
}
\description{
Read table of reference filter parameters for a single instrument
}
