% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_refraction_csv}
\alias{read_refraction_csv}
\title{Read table of indexes of refraction for cruise and population combinations}
\usage{
read_refraction_csv(path = NULL)
}
\value{
A dataframe with columns for cruise and each population's index of refraction alias
}
\description{
Read table of indexes of refraction for cruise and population combinations
}
