% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_sfl_tsv}
\alias{read_sfl_tsv}
\title{Read SFL tab-delimited file}
\usage{
read_sfl_tsv(path)
}
\arguments{
\item{path}{Path to validated tab-delimited SFL file}
}
\value{
A tibble with SFL data
}
\description{
Read SFL tab-delimited file
}
