% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{reclassify}
\alias{reclassify}
\title{Reclassify based on iformation in separate database.}
\usage{
reclassify(old_db, target_dir, mie = NULL, cores = 1, backup = TRUE)
}
\arguments{
\item{old_db}{Database file from which to pull gating parameters, metadata,
and outlier flags.}

\item{target_dir}{Directory with target database file and OPP data. There
must be a database file in target_dir with the same name as old_db. There
should also be an OPP directory with name <cruise>_opp where <cruise> is
the cruise name stored in the metadata table of old_db.}

\item{mie}{Mie theory lookup table CSV file to use in place of installed table.}

\item{cores}{Number of cores to use for VCT creation.}

\item{backup}{If backup is TRUE, create a dated backup copy of database file
and VCT directory found in target_dir before making changes.}
}
\value{
None
}
\description{
This functions automates realysis that occurs downstream from raw particle
filtration, e.g. gating, diameter, carbon. It will clear the gating, poly,
and metadata tables in the target database and copy the same tables from
old_db. It will also copy outlier flags from old_db to the target database
updated, OPP particles will be reanalyzed.
}
\examples{
\dontrun{
reclassify(old_db, target_dir, mie="mie.csv", cores=4, backup=TRUE)
}
}
