% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{remove_boundary_points}
\alias{remove_boundary_points}
\title{Remove boundary points.}
\usage{
remove_boundary_points(df, cols, max_only = FALSE)
}
\arguments{
\item{df}{Dataframe of particle data with a date column}

\item{cols}{Columns to use for boundary point definition}

\item{max_only}{Only remove maximum boundary points}
}
\value{
df with boundary points removed.
}
\description{
Boundary points are defined as points where the value for any of the
selected columns is equal to the maximum or minimum value for that column
at that date.
}
