% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{reset.db.except.params}
\alias{reset.db.except.params}
\title{Remove entries for all tables except in filter, gating, and poly tables.}
\usage{
reset.db.except.params(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
None
}
\description{
Remove entries for all tables except in filter, gating, and poly tables.
}
\examples{
\dontrun{
reset.db.except.params(db)
}
}

