% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{reset.db}
\alias{reset.db}
\title{Remove entries for all tables.}
\usage{
reset.db(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
None
}
\description{
Remove entries for all tables.
}
\examples{
\dontrun{
reset.db(db)
}
}

