% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{reset.gating.table}
\alias{reset.gating.table}
\title{Delete all rows in gating table.}
\usage{
reset.gating.table(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
None
}
\description{
Delete all rows in gating table.
}
\examples{
\dontrun{
reset.gating.table(db)
}
}
