% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{reset.opp.stats.table}
\alias{reset.opp.stats.table}
\title{Delete all rows in opp table.}
\usage{
reset.opp.stats.table(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
None
}
\description{
Delete all rows in opp table.
}
\examples{
\dontrun{
reset.opp.stats.table(db)
}
}

