% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{reset.sfl.table}
\alias{reset.sfl.table}
\title{Delete all rows in sfl table.}
\usage{
reset.sfl.table(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
None
}
\description{
Delete all rows in sfl table.
}
\examples{
\dontrun{
reset.sfl.table(db)
}
}

