% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{reset_gating_plan_table}
\alias{reset_gating_plan_table}
\title{Delete all rows in gating plan table.}
\usage{
reset_gating_plan_table(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Number of rows deleted.
}
\description{
Delete all rows in gating plan table.
}
