% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{reset_opp2_table}
\alias{reset_opp2_table}
\title{Delete all rows in opp2 table.}
\usage{
reset_opp2_table(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Number of rows deleted.
}
\description{
Delete all rows in opp2 table.
}
