% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{reset_outlier_table}
\alias{reset_outlier_table}
\title{Delete all rows in outlier table.}
\usage{
reset_outlier_table(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Number of rows deleted.
}
\description{
Delete all rows in outlier table.
}
