% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{reset_poly_table}
\alias{reset_poly_table}
\title{Delete all rows in poly table.}
\usage{
reset_poly_table(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Number of rows deleted.
}
\description{
Delete all rows in poly table.
}
