% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{reset_table}
\alias{reset_table}
\title{Delete all rows in an arbitrary SQLite3 DB table.}
\usage{
reset_table(db, table_name)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{table_name}{Table name.}
}
\value{
Number of rows deleted.
}
\description{
Delete all rows in an arbitrary SQLite3 DB table.
}
