% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save.cytdiv}
\alias{save.cytdiv}
\title{Save cytometric diversity statistics for one file to cytdiv table.}
\usage{
save.cytdiv(db, indices, cruise.name, file.name)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{indices}{List of indices N0, N1, H, J, opp_red.}

\item{cruise.name}{Cruise name.}

\item{file.name}{File name with julian day directory.}
}
\value{
None
}
\description{
Save cytometric diversity statistics for one file to cytdiv table.
}
\examples{
\dontrun{
save.cytdiv(db, indices, "testcruise", "2014_185/2014-07-04T00-00-02+00-00")
}
}

