% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save.filter.params}
\alias{save.filter.params}
\title{Save filter parameters to the filter table.}
\usage{
save.filter.params(db, params = NULL)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{params}{Named list of filter parameters. Defaults to
list(width=0.5, notch1=NA, notch2=NA, offset=0, origin=NA).}
}
\value{
None
}
\description{
Save filter parameters to the filter table.
}
\examples{
\dontrun{
save.filter.params(db)
save.fitter.params(db, list(notch1=.75, notch2=.5, offset=100, origin=NA, width=1.0))
}
}

