% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save.gating.params}
\alias{save.gating.params}
\title{Save gating parameters.}
\usage{
save.gating.params(db, poly.log)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{poly.log}{Named list of per population gating polygons.}
}
\value{
None
}
\description{
This creates an entry in the gating table and saves gating polygon
coordinates in the poly table.
}
\examples{
\dontrun{
save.gating.params(db, poly.log)
}
}

