% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save.opp.file}
\alias{save.opp.file}
\title{Save OPP as a gzipped LabView format binary file}
\usage{
save.opp.file(opp, opp.dir, file.name, untransform = FALSE)
}
\arguments{
\item{opp}{OPP data frame of filtered particles.}

\item{opp.dir}{Output directory. Julian day sub-directories will be
automatically created.}

\item{file.name}{File name with julian day directory.}

\item{untransform}{Convert linear data to log.}
}
\value{
None
}
\description{
Save OPP as a gzipped LabView format binary file
}
\examples{
\dontrun{
save.opp.file(opp, opp.dir, "2014_185/2014-07-04T00-00-02+00-00")
}
}

