% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save.outliers}
\alias{save.outliers}
\title{Save Outliers in the database}
\usage{
save.outliers(db, table.name)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{table.name}{Dataframe that contains the list of files flagged as outliers}
}
\value{
None
}
\description{
Save Outliers in the database
}
\examples{
\dontrun{
save.outliers(db,  table.name)
}
}
