% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save.poly}
\alias{save.poly}
\title{Save gating polygon coordinates in the poly table.}
\usage{
save.poly(db, poly.log, popname, gating.id)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{poly.log}{Named list of per population gating polygons.}

\item{gating.id}{Foreign key into gating table.}
}
\value{
None
}
\description{
These entries will be linked to an entry in the gating table by gating.id.
}
\examples{
\dontrun{
save.poly(db, poly.log, "d3afb1ea-ad20-46cf-866d-869300fe17f4")
}
}
