% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save.sfl}
\alias{save.sfl}
\title{Import SFL files to the database.}
\usage{
save.sfl(db, cruise, evt.dir = NULL, sfl.file = NULL)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{cruise}{Cruise name.}

\item{evt.dir}{EVT file directory with SFL files.}

\item{sfl.file}{Single SFL file to import.}
}
\value{
None
}
\description{
SFL input may come from all SFL files found in the EVT directory or from a
single SFL file. This function calls a copy of
executable_scripts/import_sfl.py installed in the package installation tree.
}
\examples{
\dontrun{
save.sfl(db, "testcruise", evt.dir=evt.dir)
save.sfl(db, "testcruise", sfl.file=sfl.file)
}
}

