% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save.vct.file}
\alias{save.vct.file}
\title{Save VCT per particle population classification.}
\usage{
save.vct.file(vct, vct.dir, file.name, quantile)
}
\arguments{
\item{vct}{List of per particle population classifications.}

\item{vct.dir}{Output directory for VCT files.}

\item{file.name}{File name with julian day directory.}

\item{quantile}{Filtering quantile for this file}
}
\value{
None
}
\description{
File will be saved to vct.dir as a gzipped text file, one line per particle.
}
\examples{
\dontrun{
save.vct.file(db, vct.dir, "2014_185/2014-07-04T00-00-02+00-00", 97.5)
}
}
