% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save.vct.stats}
\alias{save.vct.stats}
\title{Save VCT aggregate population statistics for one file to vct table.}
\usage{
save.vct.stats(db, cruise.name, file.name, opp, method, gating.id)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{cruise.name}{Cruise name.}

\item{file.name}{File name with julian day directory.}

\item{opp}{OPP data frame with pop column.}

\item{method}{Gating method name.}

\item{gating.id}{ID for entry in gating table.}
}
\value{
None
}
\description{
Save VCT aggregate population statistics for one file to vct table.
}
\examples{
\dontrun{
save.vct.stats(db, "testcruise", "2014_185/2014-07-04T00-00-02+00-00",
               opp, "Manual gating", "d3afb1ea-ad20-46cf-866d-869300fe17f4")
}
}

