% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save_filter_params_from_file}
\alias{save_filter_params_from_file}
\title{Save filter parameters from a TSV file.}
\usage{
save_filter_params_from_file(db, filter_tsv)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{filter_plan_tsv}{Filter plan TSV file.}
}
\description{
Save filter parameters from a TSV file.
}
