% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save_filter_plan}
\alias{save_filter_plan}
\title{Save filter plan to db}
\usage{
save_filter_plan(db, filter_plan)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{filter_plan}{Data Frame with two columns: start_date and filter_id.
start_date should be a UTC POSIXct date object.}
}
\value{
None
}
\description{
Save filter plan to db
}
