% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save_gating_params_from_file}
\alias{save_gating_params_from_file}
\title{Save gating parameters from a TSV file.}
\usage{
save_gating_params_from_file(db, gating_tsv)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{gating_tsv}{Gating parameters TSV file.}
}
\description{
Save gating parameters from a TSV file.
}
