% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save_gating_plan}
\alias{save_gating_plan}
\title{Save gating plan to db}
\usage{
save_gating_plan(db, gating_plan)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{gating_plan}{Data Frame with two columns: start_date and gating_id.
start_date should be a UTC POSIXct date object.}
}
\value{
None
}
\description{
Save gating plan to db
}
