% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save_metadata}
\alias{save_metadata}
\title{Save metadata to db}
\usage{
save_metadata(db, metadata)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{metadata}{One row data frame or tibble of cruise name and instrument
serial as "cruise" and "inst".}
}
\description{
Save metadata to db
}
