% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save_metadata_from_file}
\alias{save_metadata_from_file}
\title{Save metadata from an SFL file name.}
\usage{
save_metadata_from_file(db, sfl_tsv)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{sfl_tsv}{SFL TSV file.}
}
\description{
Save metadata from an SFL file name.
}
