% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save_opp2_stats}
\alias{save_opp2_stats}
\title{Save opp2}
\usage{
save_opp2_stats(db, opp2_stats)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{oppextra}{Data frame or tibble of opp2 to save. Should have columns
matching the opp2 table, except file should be named file_id. It will
be renamed to file before committing to the database.}
}
\value{
None
}
\description{
Save opp2
}
