% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save_opp_stats}
\alias{save_opp_stats}
\title{Save OPP aggregate statistics for one file/quantile combo to opp table.}
\usage{
save_opp_stats(db, opp_stats)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{opp_stats}{Data frame or tibble of OPP statistics that matches the opp
table.}
}
\value{
None
}
\description{
Save OPP aggregate statistics for one file/quantile combo to opp table.
}
