% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save_outliers}
\alias{save_outliers}
\title{Save Outliers in the database}
\usage{
save_outliers(db, outliers, overwrite = TRUE)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{outliers}{Dataframe with "file_id" and "flag" for outliers. Entries for
files already in the database will be overwritten.}

\item{overwrite}{Overwrite any existing file entries.}
}
\value{
None
}
\description{
Save Outliers in the database
}
