% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save_poly}
\alias{save_poly}
\title{Save gating polygon coordinates in the poly table.}
\usage{
save_poly(db, poly.log, popname, gating_id)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{poly.log}{Named list of per population gating polygons.}

\item{gating_id}{Foreign key into gating table.}
}
\value{
None
}
\description{
These entries will be linked to an entry in the gating table by gating_id.
}
