% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save_poly_from_file}
\alias{save_poly_from_file}
\title{Save gating polygon coordinates from a TSV file.}
\usage{
save_poly_from_file(db, poly_tsv)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{poly_tsv}{Gating polygon coordinates TSV file.}
}
\description{
Save gating polygon coordinates from a TSV file.
}
