% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save_sfl_from_file}
\alias{save_sfl_from_file}
\title{Save SFL and metadata from a TSV file.}
\usage{
save_sfl_from_file(db, sfl_tsv)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{sfl_tsv}{SFL TSV file.}
}
\description{
Save SFL and metadata from a TSV file.
}
