% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save_vct_stats}
\alias{save_vct_stats}
\title{Save VCT aggregate population statistics for one file to vct table.}
\usage{
save_vct_stats(db, vct_stats)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{vct_stats}{DataFrame of VCT statistics created by prep_vct_stats()}
}
\value{
None
}
\description{
Save VCT aggregate population statistics for one file to vct table.
}
