% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{seaflowpy_classify}
\alias{seaflowpy_classify}
\title{Classifiy particles for a list of OPP files using seaflowpy_classify.}
\usage{
seaflowpy_classify(db, cruise.name, opp.dir, vct.dir, gating.id,
  process.count = 1, limit = NULL, resolution = NULL, start.file = NULL,
  end.file = NULL)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{cruise.name}{Cruise name.}

\item{opp.dir}{OPP file directory.}

\item{vct.dir}{VCT file output directory.}

\item{gating.id}{ID for gating parameters.}

\item{process.count}{Number of processes to start for filtering.}

\item{limit}{Only process up to this many files.}

\item{resolution}{Progress update resolution in \%.}

\item{start.file}{First of subset of files to classify (including julian day dir)}

\item{end.file}{Last of subset of files to classify (including julian day dir)}

\item{opp.files}{List of OPP files to classify. Include julian day directory.}
}
\value{
None
}
\description{
Classify a  list of OPP files. Save per file aggregate population statistics
to SQLite3 database and save particle population definitions to text files
in vct.dir.
}
\examples{
\dontrun{
seaflowpy_classify(db, "testcruise", opp.dir, vct.dir,
                   "d3afb1ea-ad20-46cf-866d-869300fe17f4")
}
}

