% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{seaflowpy_filter}
\alias{seaflowpy_filter}
\title{Filter a directory of EVT files using seaflowpy_filter}
\usage{
seaflowpy_filter(db, cruise.name, evt.dir, opp.dir, process.count = 1,
  limit = NULL, resolution = NULL, origin = NULL, width = NULL,
  notch1 = NULL, notch2 = NULL, offset = NULL)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{cruise.name}{Cruise name.}

\item{evt.dir}{EVT file directory.}

\item{opp.dir}{OPP file output directory.}

\item{process.count}{Number of processes to start for filtering.}

\item{limit}{Only process up to this many files.}

\item{resolution}{Progress update resolution in \%.}

\item{origin, width, notch1, notch2, offset}{Filter parameters.}
}
\value{
None
}
\description{
Filter a list of EVT files. Save OPP per file aggregate statistics to
SQLite3 database and save particle data to binary files in opp.dir.
}
\examples{
\dontrun{
seaflowpy_filter("testcruise.db", "testcruise", "./testcruise", "./testcruise_opp")
}
}

