% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{select_files_in}
\alias{select_files_in}
\title{select_files_in}
\usage{
select_files_in(files, keep)
}
\arguments{
\item{files}{Vector of files to pull from}

\item{keep}{Vector of files to keep in <files>}
}
\value{
Subset of <files> that are in <keep> based on SeaFlow file ID
}
\description{
Only keep files in input vector if they are present in vector of files to
keep. Match based on the canonical SeaFlow file ID.
}
\examples{
\dontrun{
chosen <- select_files_in(files, keep)
}
}
