% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{set.gating.params}
\alias{set.gating.params}
\title{Define polygons for population gating.}
\usage{
set.gating.params(opp, popname, para.x, para.y, poly.log = NULL)
}
\arguments{
\item{opp}{OPP data frame.}

\item{popname}{Population name.}

\item{para.x}{Channel to use as x axis.}

\item{para.y}{Channel to use as y axis.}

\item{poly.log}{Named list of gating polygon definitions. If a definition for
popname already exists it will be updated. If it doesn"t exist it will be
appended to the end to the list. If poly.log is NULL a new list will be
created.}
}
\value{
Version of poly.log with a new polygon defintion for popname.
}
\description{
Define polygons for population gating.
}
\examples{
\dontrun{
poly.log <- set.gating.params(opp, "beads", "fsc_small", "pe")
poly.log <- set.gating.params(opp, "prochloro", "fsc_small", "chl_small",
                              poly.log)
}
}
