% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{set_gating_params}
\alias{set_gating_params}
\title{Define polygons for population gating.}
\usage{
set_gating_params(opp, popname, para_x, para_y, poly_log = NULL)
}
\arguments{
\item{opp}{OPP data frame.}

\item{popname}{Population name.}

\item{para_x}{Channel to use as x axis.}

\item{para_y}{Channel to use as y axis.}

\item{poly_log}{Named list of gating polygon definitions. If a definition for
popname already exists it will be updated. If it doesn't exist it will be
appended to the end to the list. If poly_log is NULL a new list will be
created.}
}
\value{
Version of poly.log with a new polygon defintion for popname.
}
\description{
Define polygons for population gating.
}
