% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sfl_date_where_clause}
\alias{sfl_date_where_clause}
\title{Add a WHERE clause with date conditionals to SQL statement.}
\usage{
sfl_date_where_clause(sql, start.date, end.date, append = F)
}
\arguments{
\item{start.date}{Start date in format YYYY-MM-DD HH:MM.}

\item{end.date}{End date in format YYYY-MM-DD HH:MM}

\item{db}{SQLite3 database file path.}
}
\value{
Original SQL string with new WHERE clause statements.
}
\description{
Return a copy of SQL string with a WHERE clause added to find sfl.date
entries between start.date and end.date inclusive. If append is FALSE, a new
WHERE clause is added. If append is TRUE, conditionals are added to the end
of any existing WHERE statements.
}
\examples{
\dontrun{
sql <- sfl_date_where_class(sql, "2014-07-04 00:00", "2014-07-04 00:10")
sql <- sfl_date_where_class(sql, "2014-07-04 00:00", "2014-07-04 00:10",
                            append=T)
}
}
