% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{size_carbon_conversion}
\alias{size_carbon_conversion}
\title{Estimate cell diameter and carbon cell quotas and biomass of phytoplankton populations based on normalized foward scatter to 1-µm beads used as internal standard}
\usage{
size_carbon_conversion(opp, beads_fsc, inst, mie = NULL)
}
\arguments{
\item{opp}{Table that contains fsc_small values (transformed data).}

\item{beads_fsc}{Values of the fsc_small for beads (transformed data).}

\item{inst}{Instrument serial number}

\item{mie}{Optional Mie theory lookup dataframe}
}
\value{
A dataframe with cell size and carbon cell quotas
}
\description{
Estimate cell diameter and carbon cell quotas and biomass of phytoplankton populations based on normalized foward scatter to 1-µm beads used as internal standard
}
