% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sql.dbGetQuery}
\alias{sql.dbGetQuery}
\title{Wrapper to run dbGetQuery and clean up connection on error.}
\usage{
sql.dbGetQuery(db, sql)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{sql}{SQL query to run.}
}
\value{
Data frame returned by dbGetQuery.
}
\description{
Wrapper to run dbGetQuery and clean up connection on error.
}
\examples{
\dontrun{
sql.dbGetQuery(db, "SELECT * FROM some.table")
}
}

