% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sql_dbExecute}
\alias{sql_dbExecute}
\title{Wrapper to run dbExecute and clean up connection on error.}
\usage{
sql_dbExecute(db, sql)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{sql}{SQL statement to run.}
}
\value{
Number of rows affected
}
\description{
Use for any statement except SELECT, in which case use sql_dbGetQuery.
}
