% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sql_dbGetQuery}
\alias{sql_dbGetQuery}
\title{Wrapper to run dbGetQuery and clean up connection on error.}
\usage{
sql_dbGetQuery(db, sql)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{sql}{SQL query to run.}
}
\value{
Data frame returned by dbGetQuery.
}
\description{
Use for SELECT statements only, otherwise use sql_dbExecute.
}
