% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sql_dbListTables}
\alias{sql_dbListTables}
\title{Wrapper to get list of tables in a database and clean up connection on error.}
\usage{
sql_dbListTables(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Character vector of table names.
}
\description{
Wrapper to get list of tables in a database and clean up connection on error.
}
