% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{starts_from_results_table}
\alias{starts_from_results_table}
\title{Create a tibble of start dates for gating IDs based vct / opp table entries.}
\usage{
starts_from_results_table(tbl)
}
\arguments{
\item{tbl}{VCT or OPP table as data.frame or tibble}
}
\value{
tibble with character columns start_date and gating_id
}
\description{
Create a tibble of start dates for gating IDs based vct / opp table entries.
}
