% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{stat.calibration}
\alias{stat.calibration}
\title{Calibrate abundance based on Influx data}
\usage{
stat.calibration(stat, cruisename, calib = NULL)
}
\arguments{
\item{stat}{The stat table returned from get.stat.table()}

\item{cruisename}{Name of the cruise (returned from get.mata.table()).}

\item{calib}{Optional calibration calibration  dataframe of regression values (cruise=cruise.name, pop=prochloro, a=slope, b=intercept)}
}
\value{
the stat table witrh corrected abundancce if available
}
\description{
Calibrate abundance based on Influx data
}
