% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{stat_calibration}
\alias{stat_calibration}
\title{Calibrate abundance based on Influx data}
\usage{
stat_calibration(stat, cruisename, calib = NULL)
}
\arguments{
\item{stat}{The stat table returned from get_stat_table()}

\item{cruisename}{Name of the cruise (returned from get_metadata_table()).}

\item{calib}{Optional calibration calibration  dataframe of regression values (cruise=cruise.name, pop=prochloro, a=slope, b=intercept)}
}
\value{
the stat table witrh corrected abundancce if available
}
\description{
Calibrate abundance based on Influx data
}
