% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{string.to.POSIXct}
\alias{string.to.POSIXct}
\title{Returns a POSIXct object for a human readable date string.}
\usage{
string.to.POSIXct(date.string)
}
\arguments{
\item{date.string}{Date in format YYYY-MM-DD HH:MM.}
}
\value{
POSIXct object.
}
\description{
Returns a POSIXct object for a human readable date string.
}
\examples{
\dontrun{
date.ct <- string.to.POSIXct("2014-01-24 13:03")
}
}
