% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{to_date_str}
\alias{to_date_str}
\title{Convert POSIXct date to standard SeaFlow timestamp string.}
\usage{
to_date_str(dates)
}
\arguments{
\item{dates}{Vector of POSIXct dates}
}
\value{
Timestamp in the form of 2020-07-25T14:04:05+00:00
}
\description{
Convert POSIXct date to standard SeaFlow timestamp string.
}
