% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{to_filename_date_str}
\alias{to_filename_date_str}
\title{Convert POSIXct date to standard SeaFlow filename timestamp string.}
\usage{
to_filename_date_str(dates)
}
\arguments{
\item{date}{Vector of POSIXct dates}
}
\value{
Timestamp in the form of 2020-07-25T14-04-05+00-00
}
\description{
Convert POSIXct date to standard SeaFlow filename timestamp string.
}
