% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{transform_PSD}
\alias{transform_PSD}
\title{Manipulate the size distribution created by FCSplankton::create_PSD(). 
Calculate the sum of particles in each size class over specific temporal resolution; transform the header}
\usage{
transform_PSD(
  distribution,
  time.step = "1 hour",
  Qc.to.diam = FALSE,
  interval.to.geomean = FALSE,
  abundance.to.biomass = FALSE
)
}
\arguments{
\item{distribution}{Particle size disitribution created by FCSplankton::create_PSD().
i.e., a tibble of size distribution over time. First column must be time (POSIXt class object);
Second column must name of the population; other columns represent the different size classes. 
Size classes can represent either diameter or carbon quota (assuming spherical particles).}

\item{time.step}{Time step over which to sum the number of particles in each size class. Default 1 hour, must be higher than 3 minutes}

\item{Qc.to.diam}{Convert carbon quotas to diameter as described in
Menden-Deuer, S. and Lessard, E. J. Carbon to volume relationships for dinoflagellates, diatoms, and other protist plankton.
Limnol. Oceanogr. 45, 569–579 (2000).}

\item{interval.to.geomean}{Transform size class intervals to geometric mean values 
(i.e. convert breaks (min, max] to geometric mean defined as sqrt(mean*max).}

\item{abundance.to.biomass}{Calculate carbon biomass in each population (i.e. cell abundance x Qc)}
}
\value{
Size distribution
}
\description{
Manipulate the size distribution created by FCSplankton::create_PSD(). 
Calculate the sum of particles in each size class over specific temporal resolution; transform the header
}
\examples{
\dontrun{
distribution <- transform_PSD(distribution, time.step="1 hour")
}
}
