% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{update_plan_table}
\alias{update_plan_table}
\title{Update the gating_plan or filter_plan table based on vct or opp entries}
\usage{
update_plan_table(db, results_table, check = FALSE)
}
\arguments{
\item{db}{popcycle SQLite3 database file in which to save plan table}

\item{results_table}{opp or vct table}

\item{check}{Report changes to be made without saving to the database}
}
\value{
Invisibly return the updated plan
}
\description{
Update the gating_plan or filter_plan table based on vct or opp entries
}
