% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{validate_gridded}
\alias{validate_gridded}
\title{Show messages for rows in gridded particle data that fail validation}
\usage{
validate_gridded(gridded, vct, grid_bins, refrac)
}
\arguments{
\item{gridded}{Gridded dataframe created by grid(). This function loops through
rows in this dataframe so provide a subset of a real gridded dataframe for testing.}

\item{vct}{VCT dataframe that covers time ranges present in gridded.}

\item{grid_bins}{The grid bins named list used to create gridded.}

\item{refrac}{Dataframe of population-specific refractive indices used to create gridded.}
}
\description{
Show messages for rows in gridded particle data that fail validation
}
