% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{validate_psd}
\alias{validate_psd}
\title{Show messages for rows in gridded particle data that fail validation}
\usage{
validate_psd(psd, vct, grid, refrac)
}
\arguments{
\item{psd}{Gridded dataframe created by create_PSD. This function loops through
rows in this dataframe so provide a subset of a real psd dataframe for testing.}

\item{vct}{VCT dataframe that covers time ranges present in psd.}

\item{grid}{The grid named list used to create psd.}

\item{refrac}{Dataframe of population-specific refractive indices used to create psd.}
}
\description{
Show messages for rows in gridded particle data that fail validation
}
