% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime.R
\name{write_realtime_bio_tsdata}
\alias{write_realtime_bio_tsdata}
\title{Write population data as a TSDATA file}
\usage{
write_realtime_bio_tsdata(
  bio,
  project,
  outfile,
  filetype = "SeaFlowPop",
  description = "SeaFlow population data"
)
}
\arguments{
\item{bio}{Population dataframe created by create_realtime_bio()}

\item{project}{Project identifier}

\item{outfile}{Output file path}

\item{filetype}{Filetype identifier}

\item{description}{Long form description of this file}
}
\description{
Write population data as a TSDATA file
}
