% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime.R
\name{write_realtime_meta_tsdata}
\alias{write_realtime_meta_tsdata}
\title{Write SFL metadata as a TSDATA file}
\usage{
write_realtime_meta_tsdata(
  meta,
  project,
  outfile,
  filetype = "SeaFlowSFL",
  description = "SeaFlow SFL data"
)
}
\arguments{
\item{meta}{SFL metadata dataframe created by create_realtime_meta()}

\item{project}{Project identifier}

\item{outfile}{Output file path}

\item{filetype}{Filetype identifier}

\item{description}{Long form description of this file}
}
\description{
Write SFL metadata as a TSDATA file
}
