% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load.R
\name{writeSeaflow}
\alias{writeSeaflow}
\title{Write an EVT or OPP binary file.}
\usage{
writeSeaflow(df, path, untransform = TRUE)
}
\arguments{
\item{df}{EVT or OPP data frame.}

\item{path}{Path to binary file. If this file ends with ".gz" the output
will be gzipped.}

\item{untransform}{Convert linear data to log.}
}
\value{
None
}
\description{
Write a SeaFlow LabView binary particle data file.
}
\examples{
\dontrun{
writeSeaflow(opp, "foo/2014_213/2014-07-04T00-00-02+00-00.gz")
}
}

