% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls_convert.R
\name{xls_convert}
\alias{xls_convert}
\title{Add metadata to 'stat' table and convert it to XLSX format.}
\usage{
xls_convert(db, meta, path)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{meta}{SeaFlow instrument log file.}

\item{path}{Path to save the XLSX file.}
}
\value{
none
}
\description{
Add metadata to 'stat' table and convert it to XLSX format.
}
\examples{
\dontrun{
xls_convert(db, meta, path)
}
}
