% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls_convert.R
\name{xls_merge}
\alias{xls_merge}
\title{Merge multiple 'xlsx' table and convert it to XLSX format.}
\usage{
xls_merge(list, path)
}
\arguments{
\item{list}{list of xlsx tables created by popcycle::xls_convert().}

\item{path}{Path to save the merged XLSX file.}
}
\value{
None
}
\description{
Merge multiple 'xlsx' table and convert it to XLSX format.
}
\examples{
\dontrun{
xls_merge(list, path)
}
}
