% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{add_coord_labels}
\alias{add_coord_labels}
\title{Add string coordinate labels to a dataframe of gridded SeaFlow data}
\usage{
add_coord_labels(psd, grid)
}
\arguments{
\item{psd}{Gridded dataframe created by create_PSD.}

\item{grid}{Named list of breakpoints created by create_grid().}
}
\value{
Named list identical to grid with numeric break points converted to character vectors.
}
\description{
Add string coordinate labels to a dataframe of gridded SeaFlow data
}
