% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{carbon_conversion}
\alias{carbon_conversion}
\title{Estimate Carbon cell quotas and biomass based on normalized foward scatter to 1-µm beads used as internal standard}
\usage{
carbon_conversion(phyto, inst = "740")
}
\arguments{
\item{phyto}{Table generated by the function normalization().}

\item{inst}{Serial number of the instrument}
}
\value{
A dataframe with carbon cell quotas and biomass
}
\description{
Estimate Carbon cell quotas and biomass based on normalized foward scatter to 1-µm beads used as internal standard
}
