% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{classify.opp.files}
\alias{classify.opp.files}
\title{Classify particles for a list of OPP files.}
\usage{
classify.opp.files(
  db,
  opp.dir,
  opp.files,
  vct.dir,
  gating.id = NULL,
  vct.table = NULL,
  mie.table = NULL
)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{opp.dir}{OPP file directory.}

\item{opp.files}{List of OPP files to classify. Include julian day directory.}

\item{vct.dir}{VCT file output directory.}

\item{gating.id}{Optionally provide the ID for gating parameters. If NULL,
vct.table will be used to determine gating parameters. If vct.table is null,
the latest gating parameters in the database will be used. gating.id takes
precedence over vct.table.}

\item{vct.table}{Optionally provide a previous VCT table that will be used to
gating parameters by date range. If NULL, gating.id will be used for all
files. If gating.id is NULL, the latest gating parameters in the database
will be used.}

\item{mie.table}{Optionally provide the Mie theory table. If NULL,
the installed Mie Theory lookup table will be used. This table should be
identical to one produced by read_mie_csv().}
}
\value{
None
}
\description{
Classify a  list of OPP files. Save per file aggregate population statistics
to SQLite3 database and save particle population definitions to text files
in vct.dir.
}
\examples{
\dontrun{
classify.opp.files(db, opp.dir, opp.files, vct.dir)
classify.opp.files(db, opp.dir, opp.files, vct.dir,
                   "d3afb1ea-ad20-46cf-866d-869300fe17f4")
}
}
