% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{classify.opp}
\alias{classify.opp}
\title{Classify particles from an OPP dataframe.}
\usage{
classify.opp(opp, gates.log)
}
\arguments{
\item{opp}{SQLite3 database file path.}

\item{gates.log}{A gating scheme from the function 'add.manual.classification()' or 'add.auto.classification()'}
}
\value{
List of per particle classifications
}
\description{
Classify particles from an OPP dataframe using a gating scheme provided by gates.log.
}
\examples{
\dontrun{
opp <- classify.opp(opp, gates.log)
}
}
