% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{concatenate.evtopp}
\alias{concatenate.evtopp}
\title{Concatenate EVT or OPP files}
\usage{
concatenate.evtopp(evtopp.list, n = 1e+05, min.fsc = 0, min.pe = 0,
  min.chl = 0, transform = TRUE, ...)
}
\arguments{
\item{evtopp.list}{List of EVT or OPP files (full path required).}

\item{n}{Number of rows to return.}

\item{min.fsc, }{min.pe, min.chl Minimum value for fsc_small, pe and chl_small respectively}
}
\value{
A dataframe with n rows.
}
\description{
Concatenate EVT or OPP files
}
