% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{create_breaks}
\alias{create_breaks}
\title{Create breaks vector that can be used with cut().}
\usage{
create_breaks(bin_count, min_val, max_val, log_base = NULL, log_answers = TRUE)
}
\arguments{
\item{bin_count}{Number of bins between breaks.}

\item{min_val}{Value for the minimum break (fencepost).}

\item{max_val}{Value for the maximum break (fencepost).}

\item{log_base}{Log base to use when creating log-spaced bins for distribution.
If NULL the breaks will not be log-spaced.}

\item{log_answers}{If log_base is provided, this determines whether the break
values themselves are log values or not. e.g. if TRUE breaks may be
[1, 2, 3], if FALSE breaks may be [10, 100, 1000].}
}
\value{
Vector defining breaks for use with cut(), length == bin_count + 1
}
\description{
Create breaks vector that can be used with cut().
}
