% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{create_PSD}
\alias{create_PSD}
\title{Create particle size distribution}
\usage{
create_PSD(db, vct.dir, breaks, quantile = 50)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{vct.dir}{VCT file directory..}

\item{breaks}{Breaks must be a vector of values defining the breaks for the size distribution.}

\item{quantile}{OPP Filtering quantile.}
}
\value{
Size distribution
}
\description{
Create particle size distribution
}
\examples{
\dontrun{
 
breaks <- # for Qc
min <- 0.003 # minimum quotas (3 fgC / cell)
delta <- 1/18 # to define the width of each bin
m <- 300 # number of bins
breaks <- round(min*2^(((1:m)-1)*delta),4) # log2 space bin
print(breaks)

distribution <- create_PSD(db, vct.dir, breaks, quantile = 50)
}
}
