% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime.R
\name{create_realtime_bio}
\alias{create_realtime_bio}
\title{Create a population data tibble for one quantile from the VCT}
\usage{
create_realtime_bio(db, quantile, correction = NULL, with_abundance = FALSE)
}
\arguments{
\item{db}{popcycle database file.}

\item{quantile}{OPP filtering quantile to use.}

\item{with_abundance}{Include volume normalized "abundance" column}
}
\value{
A tibble of realtime population data
}
\description{
Create a population data tibble for one quantile from the VCT
}
