% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime.R
\name{create_realtime_meta}
\alias{create_realtime_meta}
\title{Create a metadata tibble for one quantile appopriate for realtime analysis}
\usage{
create_realtime_meta(db, volume = NULL)
}
\arguments{
\item{db}{popcycle database file.}

\item{volume}{Use a constant volume value, overriding any calculated values.}

\item{quantile}{OPP filtering quantile to use.}
}
\value{
A tibble of realtime SFL and OPP table data
}
\description{
Create a metadata tibble for one quantile appopriate for realtime analysis
}
