% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{create_volume_table}
\alias{create_volume_table}
\title{Create a tibble of volumes that can be used to calculate SeaFlow abundances}
\usage{
create_volume_table(meta, time_expr = "1 hour", median_opp_evt_ratio = NULL)
}
\arguments{
\item{meta}{Metadata dataframe created by create_meta()}

\item{time_expr}{Time expression passed to lubridate::floor_date to group datetimes
before calculating volumes. If NULL volumes are returned at their original time
resolution.}

\item{median_opp_evt_ratio}{Fixed ratio to use.}
}
\value{
A tibble of sample and virtual core volumes.
}
\description{
Create a tibble of volumes that can be used to calculate SeaFlow abundances
}
