% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter.notch}
\alias{filter.notch}
\title{Filter EVT particles.}
\usage{
filter.notch(evt, filter.params)
}
\arguments{
\item{evt}{EVT data frame.}

\item{filter.params}{Filtering parameters in a one row data frame or named
list. Columns should include width, notch.small.D1, notch.small.D2,
notch.large.D1, notch.large.D2, offset.small.D1, offset.small.D2,
offset.large.D1, offset.large.D2.}
}
\value{
Named list where list$params contains a list of filtering parameters
  and list$opp contains OPP data frame.
}
\description{
Filter EVT particles.
}
\examples{
\dontrun{
filt <- filter.notch(evt, params)
}
}
