% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter.notch}
\alias{filter.notch}
\title{Filter EVT particles.}
\usage{
filter.notch(evt, origin = NA, width = 1, notch1 = NA, notch2 = NA,
  offset = 0)
}
\arguments{
\item{evt}{EVT data frame.}

\item{origin, width, notch1, notch2, offset}{Filtering parameters. origin,
notch1, and notch2 will be calculated if NA.}
}
\value{
Named list where list$params contains a list of filtering parameters
  and list$opp contains OPP data frame.
}
\description{
Filter EVT particles.
}
\examples{
\dontrun{
filt <- filter.notch(evt, origin=NA, width=0.5, notch1=NA, notch2=NA, offset=0)
}
}

