% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{flowrate}
\alias{flowrate}
\title{Calculation of flow rate based on stream pressure measured by SeaFlow}
\usage{
flowrate(stat, inst = NULL)
}
\arguments{
\item{inst}{Instrument serial. If not provided this will attempt to be
parsed from the SFL file name (<cruise>_<serial>.sfl).}

\item{dataframe}{that contains stream pressure measured by SeaFlow (available in SFL table and STAT table). Note that the dataframe needs to have a column named 'stream_pressure'}
}
\value{
A dataframe with flow rate estimates
}
\description{
Calculation of flow rate based on stream pressure measured by SeaFlow
}
