% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get.clean.stat.table}
\alias{get.clean.stat.table}
\title{Get aggregate statistics data frame along with propagation of errors}
\usage{
get.clean.stat.table(db, inst = NULL)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{inst}{Instrument serial. If not provided this will attempt to be
parsed from the SFL file name (<cruise>_<serial>.sfl).}
}
\value{
Data frame of aggregate statistics.
}
\description{
Get aggregate statistics data frame along with propagation of errors
}
