% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.gating.params.by.id}
\alias{get.gating.params.by.id}
\title{Get gating parameters by ID.}
\usage{
get.gating.params.by.id(db, gating.id)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{gating.id}{ID in gating table and poly table.}
}
\value{
Named list where list$gates.log a recreation of original gating
  created by add.manual.classification(), add.auto.classification(), and
  save.gating.params()
}
\description{
Get gating parameters by ID.
}
\examples{
\dontrun{
gating.params <- get.gating.params.by.id(db, "d3afb1ea-ad20-46cf-866d-869300fe17f4")
}
}
