% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.gating.params.latest}
\alias{get.gating.params.latest}
\title{Get the latest gating parameters.}
\usage{
get.gating.params.latest(db)
}
\arguments{
\item{db}{SQLite3 database file path.}
}
\value{
Named list where list$row is gating entry in gating table and
  list$poly.log is list of polygon coordinates for each population.
}
\description{
Get the latest gating parameters.
}
\examples{
\dontrun{
gating.params <- get.gating.params.latest(db)
}
}

