% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get.opp.by.date}
\alias{get.opp.by.date}
\title{Get OPP data frame by date and quantile}
\usage{
get.opp.by.date(db, opp.dir, quantile, start.date, end.date,
  channel = NULL, transform = TRUE, vct.dir = NULL, pop = NULL)
}
\arguments{
\item{db}{SQLite3 data file path.}

\item{opp.dir}{OPP file directory.}

\item{quantile}{Filtering quantile for this file}

\item{start.date}{Start date in format YYYY-MM-DD HH:MM.}

\item{end.date}{End date in format YYYY-MM-DD HH:MM.}

\item{channel}{Channels to keep in returned data frame. Can be a single name
or a vector. Choosing fewer channels can significantly speed up retrieval.}

\item{transform}{Linearize OPP data.}

\item{vct.dir}{VCT file directory. If not specified returned data frame will
not have a pop column.}

\item{pop}{If specified, the returned data frame will only contain entries
for this population.}
}
\value{
Data frame of OPP data for all files between start.date and end.date
  inclusive.
}
\description{
Requires a populated sfl table.
}
\examples{
\dontrun{
opp <- get.opp.by.date(db, opp.dir, "2014-07-04 00:00", "2014-07-04 00:10")
opp <- get.opp.by.date(db, opp.dir, "2014-07-04 00:00", "2014-07-04 00:10",
                       channel=c("fsc_small", "chl_small", "pe"),
                       transform=F, vct.dir=vct.dir)
}
}
