% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get.stat.table}
\alias{get.stat.table}
\title{Get aggregate statistics data frame along with estimates of cell abundance.}
\usage{
get.stat.table(db, inst = NULL)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{inst}{Instrument serial. If not provided this will attempt to be
parsed from the SFL file name (<cruise>_<serial>.sfl).}
}
\value{
Data frame of aggregate statistics.
}
\description{
Get aggregate statistics data frame along with estimates of cell abundance.
}
\examples{
\dontrun{
stats <- get.stat.table(db, inst=NULL)
}
}
