% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get_clean_stat_table}
\alias{get_clean_stat_table}
\title{Clean the stat table by selecting the best refractive index from the table (RefracIndices_percruise.csv),
keep only the clean data (flag ==0 ) and remove percentile column}
\usage{
get_clean_stat_table(db, refracs = NULL)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{refracs}{Dataframe of per-population refractive indices. If NULL, will
read the table from this package.}
}
\value{
A cleaned stat table with corrected refractive index
}
\description{
Clean the stat table by selecting the best refractive index from the table (RefracIndices_percruise.csv),
keep only the clean data (flag ==0 ) and remove percentile column
}
