% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{get_vct_range}
\alias{get_vct_range}
\title{Find min/max for a data column / quantile pair in VCT files}
\usage{
get_vct_range(
  vct_files,
  data_cols,
  quantile,
  pop = NULL,
  ignore_dates = NULL,
  refracs = NULL,
  cores = 1
)
}
\arguments{
\item{vct_files}{Vector of VCT file paths.}

\item{data_cols}{Character vector of VCT columns to use as size metric:
fsc_small, chl_small, pe, Qc, or diam.
For Qc and diam, refractive index alias and quantile will be added to the
column name automatically. If refracs is provided, the appropriate refractive
index will automatically be applied to Qc and diam and only Qc and diam will
be returned as column names if requested. i.e. there will be no refractive
index alias suffix such as "_lwr".}

\item{quantile}{OPP Filtering quantile.}

\item{pop}{Single population label to filter for.}

\item{ignore_dates}{Don't process VCT data with these dates.}

\item{refracs}{Dataframe of refractive indices to use for each population. Column names
should be population names used in classification, values should be mid, lwr, upr.
If NULL all refractive indexes will be used.}

\item{cores}{Number of cores to use}
}
\value{
Two item numeric vector of c(min_val, max_val)
}
\description{
Find min/max for a data column / quantile pair in VCT files
}
