% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{get_vct_range}
\alias{get_vct_range}
\title{Find min/max for a data column / quantile pair in many VCT directories}
\usage{
get_vct_range(vct_dirs, data_col, quantile, cores = 1)
}
\arguments{
\item{vct_dirs}{Vector of VCT file directory paths.}

\item{data_col}{VCT column to use as size metric: fsc_small, chl_small, pe, Qc, or diam. If
Qc or diam is selected, the refractive index is indicated in the extra column "n".}

\item{quantile}{OPP Filtering quantile.}

\item{cores}{Number of worker processes to create.}
}
\value{
Two item numeric vector of c(min_val, max_val)
}
\description{
Find min/max for a data column / quantile pair in many VCT directories
}
\examples{
\dontrun{
minmax <- get_vct_range(vct_dirs, "fsc_small", quantile = 50)
}
}
