% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{get_vct_range_one_file}
\alias{get_vct_range_one_file}
\title{Find min/max for a data column / quantile pair in one VCT file}
\usage{
get_vct_range_one_file(vct_file, data_col, quantile)
}
\arguments{
\item{vct_file}{VCT file path.}

\item{data_col}{VCT column to use as size metric: fsc_small, chl_small, pe, Qc, or diam. If
Qc or diam is selected, the refractive index is indicated in the extra column "n".}

\item{quantile}{OPP Filtering quantile.}
}
\value{
Two item numeric vector of c(min_val, max_val)
}
\description{
Find min/max for a data column / quantile pair in one VCT file
}
\examples{
\dontrun{
minmax <- get_vct_range_one_file(vct_file, "fsc_small", quantile = 50)
}
}
