% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{normalization}
\alias{normalization}
\title{Normalization by beads used as internal standard}
\usage{
normalization(stat, channel = "fsc_small", norm = NULL)
}
\arguments{
\item{stat}{Stat table generated by the function get.stat.table().}

\item{channel}{Channel (fsc_small by default) to be normalized by beads (used as internal standard).}

\item{norm}{Median and SD value of beads signal used for the normalization. If NULL (default), median and sd value of beads of the lower quantile from the Stat table will be used as input.}
}
\value{
A dataframe with normalized channel values
}
\description{
Normalization by beads used as internal standard
}
