% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{normalization}
\alias{normalization}
\title{Normalization by beads used as internal standard}
\usage{
normalization(stat, spar = 0.5, plot.beads = T)
}
\arguments{
\item{stat}{Stat table generated by the function get.stat.table().}

\item{spar}{smooothing parameter, the higher the more smoothing is applied.}

\item{plot.beads}{Do you want to plot beads signal over time for fsc_small, chl_small and pe? (default is TRUE)}
}
\value{
A stat dataframe with normalized fsc_small, chl_small and pe values
}
\description{
Normalization by beads used as internal standard
}
