% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{plot.filter.cytogram.by.file}
\alias{plot.filter.cytogram.by.file}
\title{Plot helpful cytograms for exploring filtering parameters for one file.}
\usage{
plot.filter.cytogram.by.file(evt.dir, file.name, origin = NA, width = 0.5,
  notch1 = NA, notch2 = NA, offset = 0)
}
\arguments{
\item{evt.dir}{EVT file directory.}

\item{file.name}{File name with julian day directory.}

\item{origin, width, notch1, notch2, offset}{Filtering parameters. origin,
notch1, and notch2 will be calculated if NA.}
}
\value{
None
}
\description{
Plot helpful cytograms for exploring filtering parameters for one file.
}

