% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{plot.filter.cytogram}
\alias{plot.filter.cytogram}
\title{Plot helpful cytograms for exploring filtering parameters.}
\usage{
plot.filter.cytogram(evt, origin = NA, width = 1, notch1 = NA,
  notch2 = NA, offset = 0)
}
\arguments{
\item{evt}{EVT data frame.}

\item{origin, width, notch1, notch2, offset}{Filtering parameters. origin,
notch1, and notch2 will be calculated if NA.}
}
\value{
None
}
\description{
Plot helpful cytograms for exploring filtering parameters.
}

