% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.map}
\alias{plot.map}
\title{Plot cell abundances of a population on a map.}
\usage{
plot.map(stat, popname, param, ...)
}
\value{
None
}
\description{
Plot cell abundances of a population on a map.
}

