% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.vct.cytogram}
\alias{plot.vct.cytogram}
\title{Plot cytogram with particles colored by population.}
\usage{
plot.vct.cytogram(opp, para.x = "fsc_small", para.y = "chl_small",
  min.count = 300, ...)
}
\arguments{
\item{opp}{OPP data frame.}

\item{para.x}{Channel to use as x axis.}

\item{para.y}{Channel to use as y axis.}

\item{min.count}{Minimum number of particles for drawing contour lines.}

\item{...}{Additional parameters for plot()}
}
\value{
None
}
\description{
Plot cytogram with particles colored by population.
}
