% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_cruisetrack}
\alias{plot_cruisetrack}
\title{Plot cruise track on a map.}
\usage{
plot_cruisetrack(stat, param)
}
\arguments{
\item{sfl}{sfl table from get.sfl.table function}

\item{param}{Parameter to display}

\item{bin}{An integer representing the time interval (in hour) into which time is to be averaged}
}
\value{
None
}
\description{
Plot cruise track on a map.
}
