% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_filter_cytogram}
\alias{plot_filter_cytogram}
\title{Plot cytograms for exploring filtering parameters.}
\usage{
plot_filter_cytogram(evt, filter_params)
}
\arguments{
\item{evt}{EVT data frame.}

\item{filter_params}{Filtering parameters for all quantiles. Columns should
include width, notch.small.D1, notch.small.D2,
notch.large.D1, notch.large.D2, offset.small.D1, offset.small.D2,
offset.large.D1, offset.large.D2.}

\item{quantile}{Quantile to use when plotting, default = 2.5.}
}
\value{
None
}
\description{
Plot cytograms for exploring filtering parameters.
}
