% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_time}
\alias{plot_time}
\title{plot.time}
\usage{
plot.time(stat, popname,param)
}
\arguments{
\item{stat}{Stat table from get.stat.table function}

\item{param}{Parameter to display}

\item{transform}{Log transformation of the parameter'}
}
\value{
None
}
\description{
plot.time
}
