% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save.filter.params}
\alias{save.filter.params}
\title{Save filter parameters to the filter table.}
\usage{
save.filter.params(db, filter.params)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{filter.params}{Data frame of filtering parameters one row per
quantile. Columns should include:
quantile, beads.fsc.small, beads.D1, beads.D2, width,
notch.small.D1, notch.small.D2, notch.large.D1, notch.large.D2,
offset.small.D1, offset.small.D2, offset.large.D1, offset.large.D2.}

\item{beads.fsc.small}{Small forward scatter of 1µm beads used to determine
filter.params}

\item{beads.D1}{D2 of 1µm beads used to determine filter.params}
}
\value{
Database filter ID string.
}
\description{
Save filter parameters to the filter table.
}
\examples{
\dontrun{
filter.id <- save.fitter.params(db, 1000, 2000, 3000, filter.params)
}
}
