% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{save.sfl}
\alias{save.sfl}
\title{Import SFL files to the database.}
\usage{
save.sfl(db, sfl.file, cruise = NULL, inst = NULL)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{sfl.file}{Single SFL file to import.}

\item{cruise}{Cruise name. If not provided this will attempt to be
parsed from the SFL file name (<cruise>_<serial>.sfl).}

\item{inst}{Instrument serial. If not provided this will attempt to be
parsed from the SFL file name (<cruise>_<serial>.sfl).}
}
\value{
None
}
\description{
This function calls seaflowpy sfl db.
}
\examples{
\dontrun{
save.sfl(db, sfl.file=sfl.file)
}
}
