% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{seaflowpy_filter}
\alias{seaflowpy_filter}
\title{Filter a directory of EVT files using seaflowpy_filter}
\usage{
seaflowpy_filter(db, cruise.name, evt.dir, opp.dir, process.count = 1,
  twopass = FALSE, limit = NULL, resolution = NULL, width = NULL,
  notch.small.D1 = NULL, notch.small.D2 = NULL, notch.large.D1 = NULL,
  notch.large.D2 = NULL, offset.small.D1 = NULL, offset.small.D2 = NULL,
  offset.large.D1 = NULL, offset.large.D2 = NULL)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{cruise.name}{Cruise name.}

\item{evt.dir}{EVT file directory.}

\item{opp.dir}{OPP file output directory.}

\item{process.count}{Number of processes to start for filtering.}

\item{limit}{Only process up to this many files.}

\item{resolution}{Progress update resolution in \%.}

\item{width, notch.small.D1, }{notch.small.D2, notch.large.D1, notch.large.D2, offset.small.D1, offset.small.D2, offset.large.D1, offset.large.D2 Filter parameters.}
}
\value{
None
}
\description{
Filter a list of EVT files. Save OPP per file aggregate statistics to
SQLite3 database and save particle data to binary files in opp.dir.
}
\examples{
\dontrun{
seaflowpy_filter("testcruise.db", "testcruise", "./testcruise", "./testcruise_opp")
}
}
