% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{size.carbon.conversion}
\alias{size.carbon.conversion}
\title{Estimate cell diameter and carbon cell quotas and biomass of phytoplankton populations based on normalized foward scatter to 1-µm beads used as internal standard}
\usage{
size.carbon.conversion(stat, inst = NULL)
}
\arguments{
\item{stat}{Table that contains fsc_small values normalized by 1 micron beads.}

\item{inst}{Instrument serial. If not provided this will attempt to be
parsed from the SFL file name (<cruise>_<serial>.sfl).}
}
\value{
A dataframe with cell size, carbon cell quotas and biomass
}
\description{
Estimate cell diameter and carbon cell quotas and biomass of phytoplankton populations based on normalized foward scatter to 1-µm beads used as internal standard
}
