% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sql.dbWriteTable}
\alias{sql.dbWriteTable}
\title{Wrapper to run dbWriteTable and clean up connection on error.}
\usage{
sql.dbWriteTable(db, name, value)
}
\arguments{
\item{db}{SQLite3 database file path.}

\item{name}{Table name.}

\item{value}{Data frame to write.}
}
\description{
Wrapper to run dbWriteTable and clean up connection on error.
}
\examples{
\dontrun{
sql.dbWriteTable(db, name="vct", value=df)
}
}
